refactor(frontend): apply Prism Console visual redesign to chat, tasks, and projects pages
- Redesign ChatInputBar with glass-morphism backdrop-blur + Prism tokens
- Style chat message bubbles with role-specific accent colours
- Apply Prism primary/soft/border to task list cards and action buttons
- Add glass header to TaskHeaderBar and ProjectCanvas toolbar
- Theme React Flow edges, handles, minimap with Prism palette
- Update ProjectSidebar and ProjectDropZone with Prism selection states
- Split i18n messages.ts (2159→21 lines) into en/zh feature modules with catalog.ts
- Update React Flow CSS variables to use Prism tokens
