+++ frontend/src/components/chat/ChatInputBar.tsx
@@ -59,13 +59,13 @@ export default function ChatInputBar({
   }, []);
 
   return (
-    <div className="pointer-events-none bg-gradient-to-t from-white/30 dark:from-[#1f1f1f]/30 to-transparent pt-10 pb-6 z-10 transition-colors duration-300">
+    <div className="pointer-events-none bg-gradient-to-t from-[var(--surface)]/60 to-transparent pt-10 pb-6 z-10 transition-colors duration-300">
       <div className="relative max-w-[760px] mx-auto px-4 pointer-events-auto">
         {scrollButtonVisible && onScrollToBottom && (
           <button
             type="button"
             onClick={onScrollToBottom}
-            className="absolute -top-12 left-1/2 -translate-x-1/2 bg-white/50 dark:bg-[#2a2a2a]/50 border border-white/60 dark:border-white/10 w-8 h-8 rounded-full flex items-center justify-center text-gray-500 dark:text-gray-400 hover:text-gray-800 dark:hover:text-gray-200 hover:bg-white/70 dark:hover:bg-[#333]/70 shadow-sm transition-all animate-in fade-in slide-in-from-bottom-2 z-20 cursor-pointer backdrop-blur-[12px]"
+            className="absolute -top-12 left-1/2 -translate-x-1/2 bg-[var(--prism-glass)] border border-[var(--border)] w-8 h-8 rounded-full flex items-center justify-center text-[var(--text-secondary)] hover:text-[var(--text)] hover:bg-[var(--surface)] shadow-[var(--shadow-sm)] transition-all z-20 cursor-pointer backdrop-blur-xl"
             aria-label={t('chat.scrollToBottom')}
           >
             <svg
@@ -82,10 +82,10 @@ export default function ChatInputBar({
           </button>
         )}
 
-        <div className="relative bg-white/40 dark:bg-[#2a2a2a]/40 backdrop-blur-[32px] backdrop-saturate-[180%] border border-white/60 dark:border-white/10 focus-within:border-white/80 dark:focus-within:border-white/20 focus-within:bg-white/50 dark:focus-within:bg-[#333]/50 rounded-[28px] transition-all duration-300 shadow-[0_8px_32px_rgba(0,0,0,0.06),inset_0_1px_2px_rgba(255,255,255,0.8)] dark:shadow-[0_8px_32px_rgba(0,0,0,0.2),inset_0_1px_2px_rgba(255,255,255,0.05)] focus-within:shadow-[0_12px_48px_rgba(0,0,0,0.1),inset_0_1px_2px_rgba(255,255,255,1)] dark:focus-within:shadow-[0_12px_48px_rgba(0,0,0,0.3),inset_0_1px_2px_rgba(255,255,255,0.1)]">
+        <div className="relative bg-[var(--prism-glass)] backdrop-blur-xl backdrop-saturate-[180%] border border-[var(--border-strong)] focus-within:border-[var(--prism-primary-border)] focus-within:bg-[var(--surface)]/90 rounded-[28px] transition-all duration-300 shadow-[var(--shadow-input)] focus-within:shadow-[var(--shadow-pane)]">
           <textarea
             ref={textareaRef}
-            className="w-full bg-transparent border-none outline-none focus-visible:outline-none resize-none py-4 px-6 pr-[60px] pb-[46px] text-sm placeholder-gray-400 dark:placeholder-gray-500 text-[var(--text)] block leading-relaxed max-h-[200px]"
+            className="w-full bg-transparent border-none outline-none focus-visible:outline-none resize-none py-4 px-6 pr-[60px] pb-[46px] text-sm placeholder:text-[var(--text-tertiary)] text-[var(--text)] block leading-relaxed max-h-[200px]"
             placeholder={t('chat.inputPlaceholder')}
             rows={1}
             value={value}
@@ -104,8 +104,8 @@ export default function ChatInputBar({
               disabled={disabled || !value.trim()}
               className={`w-8 h-8 rounded-full flex items-center justify-center transition-all ${
                 value.trim() && !disabled
-                  ? 'bg-[var(--text)] text-[var(--background)] hover:opacity-80 hover:scale-105 active:scale-95 cursor-pointer shadow-md'
-                  : 'bg-black/5 dark:bg-white/5 text-black/20 dark:text-white/20 cursor-not-allowed'
+                  ? 'bg-[var(--prism-primary)] text-white hover:bg-[var(--prism-primary-hover)] hover:scale-105 active:scale-95 cursor-pointer shadow-[var(--shadow-sm)]'
+                  : 'bg-[var(--bg-secondary)] text-[var(--text-tertiary)] cursor-not-allowed'
               }`}
               aria-label={t('chat.send')}
             >
@@ -113,24 +113,24 @@ export default function ChatInputBar({
             </button>
           </div>
 
-          <div className="absolute left-4 bottom-2.5 flex items-center gap-2 text-[var(--text-secondary)]">
+          <div className="absolute left-4 bottom-2.5 flex items-center gap-1 text-[var(--text-tertiary)]">
             <button
               type="button"
-              className="p-1.5 hover:bg-[var(--bg-secondary)]/50 rounded-lg transition-colors"
+              className="p-1.5 hover:bg-[var(--prism-primary-soft)] hover:text-[var(--prism-primary)] rounded-lg transition-colors"
               aria-label={t('chat.attach')}
             >
-              <Plus className="w-5 h-5" strokeWidth={2} />
+              <Plus className="w-5 h-5" strokeWidth={1.8} />
             </button>
             <button
               type="button"
-              className="p-1.5 hover:bg-[var(--bg-secondary)]/50 rounded-lg transition-colors"
+              className="p-1.5 hover:bg-[var(--prism-primary-soft)] hover:text-[var(--prism-primary)] rounded-lg transition-colors"
               aria-label={t('chat.webSearch')}
             >
-              <Globe className="w-5 h-5" strokeWidth={2} />
+              <Globe className="w-5 h-5" strokeWidth={1.8} />
             </button>
             <button
               type="button"
-              className="p-1.5 hover:bg-[var(--bg-secondary)]/50 rounded-lg transition-colors"
+              className="p-1.5 hover:bg-[var(--prism-primary-soft)] hover:text-[var(--prism-primary)] rounded-lg transition-colors"
               aria-label={t('chat.reason')}
             >
               <ReasonIcon className="w-5 h-5" />
@@ -138,7 +138,7 @@ export default function ChatInputBar({
           </div>
         </div>
 
-        <div className="text-center text-[11px] text-[var(--text-secondary)]/80 font-medium mt-3 transition-colors">
+        <div className="text-center text-[11px] text-[var(--text-tertiary)] font-medium mt-3 transition-colors">
           {t('chat.disclaimer')}
         </div>
       </div>
