+++ frontend/src/components/chat/ChatMessageList.tsx
@@ -120,7 +120,7 @@ export default function ChatMessageList({
         <button
           type="button"
           onClick={scrollToBottom}
-          className="absolute bottom-36 left-1/2 -translate-x-1/2 bg-[var(--surface)] border border-[var(--border)] w-8 h-8 rounded-full flex items-center justify-center text-[var(--text-secondary)] hover:text-[var(--text)] hover:bg-[var(--bg-secondary)] shadow-sm transition-all animate-in fade-in slide-in-from-bottom-2 z-20 cursor-pointer"
+          className="absolute bottom-36 left-1/2 -translate-x-1/2 bg-[var(--prism-glass)] backdrop-blur-xl border border-[var(--border)] w-8 h-8 rounded-full flex items-center justify-center text-[var(--text-secondary)] hover:text-[var(--text)] hover:bg-[var(--surface)] shadow-[var(--shadow-sm)] transition-all z-20 cursor-pointer"
           aria-label={t('chat.scrollToBottom')}
         >
           <svg
