+++ frontend/src/components/chat/ChatThinkingBlock.tsx
@@ -11,7 +11,7 @@ export default function ChatThinkingBlock({ content }: ChatThinkingBlockProps) {
   const [isExpanded, setIsExpanded] = useState(false);
 
   return (
-    <div className="border-l-2 border-[var(--border)] pl-4 py-1 flex flex-col gap-2 transition-colors">
+    <div className="border-l-2 border-[var(--color-msg-thinking)]/30 pl-4 py-1 flex flex-col gap-2 transition-colors">
       <button
         type="button"
         className="flex items-center gap-2 cursor-pointer select-none w-fit"
