+++ frontend/src/components/project/ProjectDropZone.tsx
@@ -25,12 +25,12 @@ export default function ProjectDropZone({
 
   return (
     <div
-      className={`pointer-events-auto absolute left-0 top-0 bottom-0 z-50 w-56 overflow-y-auto border-r border-[var(--border)] bg-[var(--surface)]/95 p-3 shadow-2xl backdrop-blur-sm transition-transform duration-300 ease-out ${
+      className={`pointer-events-auto absolute left-0 top-0 bottom-0 z-50 w-56 overflow-y-auto border-r border-[var(--border)] bg-[var(--prism-glass)] backdrop-blur-xl p-3 shadow-[var(--shadow-pane)] transition-transform duration-300 ease-out ${
         visible ? 'translate-x-0' : '-translate-x-full'
       }`}
       aria-hidden={!visible}
     >
-      <p className="mb-2 px-1 text-[11px] font-medium uppercase tracking-wide text-[var(--text-secondary)]">
+      <p className="mb-2 px-1 text-[11px] font-semibold uppercase tracking-widest text-[var(--text-tertiary)]">
         {t('pages.projects.moveToProject')}
       </p>
       <div className="flex flex-col gap-2">
@@ -39,12 +39,12 @@ export default function ProjectDropZone({
             key={project.project_id}
             data-project-id={project.project_id}
             style={{ transitionDelay: `${i * 40}ms` }}
-            className={`rounded-lg border-2 bg-[var(--bg)] p-3 transition-all duration-200 hover:scale-[1.03] hover:border-[var(--apple-blue)] hover:shadow-md ${
+            className={`rounded-lg border-2 bg-[var(--bg)] p-3 transition-all duration-200 hover:scale-[1.03] hover:border-[var(--prism-primary)] hover:shadow-md ${
               visible ? 'translate-y-0 opacity-100' : 'translate-y-1 opacity-0'
             }`}
           >
             <div className="flex items-center gap-2">
-              <Folder size={15} className="shrink-0 text-[var(--apple-blue)]" />
+              <Folder size={15} className="shrink-0 text-[var(--prism-primary)]" />
               <div className="min-w-0">
                 <div className="truncate text-sm font-medium text-[var(--text)]">{project.name}</div>
                 {project.description ? (
