+++ frontend/src/shared/i18n/en/auth.ts
@@ -1,0 +1,29 @@
+// auth messages — English
+const auth = {
+    auth: {
+      login: 'Log in',
+      logout: 'Log out',
+      username: 'Username',
+      usernameHint: 'letters, digits, dot, underscore, hyphen only',
+      password: 'Password',
+      displayName: 'Display Name',
+      confirmPassword: 'Confirm Password',
+      passwordMismatch: 'Passwords do not match',
+      registerFailed: 'Registration failed',
+      registrationSubmitted: 'Registration Submitted',
+      backToLogin: 'Back to Login',
+      registerLink: "Don't have an account? Register",
+      loginLink: 'Already have an account? Log in',
+      register: 'Register',
+      pendingApproval: 'Your account is pending admin approval.',
+      changePassword: 'Change Password',
+      mustChangePassword: 'You must change your password before continuing.',
+      currentPassword: 'Current Password',
+      newPassword: 'New Password',
+      changePasswordFailed: 'Failed to change password',
+      passwordTooShort: 'Password must be at least 4 characters',
+      loginFailed: 'Login failed',
+    },
+};
+
+export default auth;
