+++ frontend/src/shared/i18n/en/common.ts
@@ -1,0 +1,105 @@
+// common messages — English
+const common = {
+    chat: {
+      thinking: 'Thinking process',
+      disclaimer: 'AI can make mistakes. Please verify important information.',
+      inputPlaceholder: 'Message AI...',
+      send: 'Send message',
+      attach: 'Attach file',
+      webSearch: 'Web search',
+      reason: 'Reason',
+      copy: 'Copy',
+      retry: 'Retry',
+      collapseTools: 'Collapse tools',
+      scrollToBottom: 'Scroll to bottom',
+    },
+    common: {
+      appName: 'AINRF Console',
+      language: 'Language',
+      english: 'English',
+      chinese: '中文',
+      loading: 'Loading...',
+      saving: 'Saving…',
+      reset: 'Reset',
+      cancel: 'Cancel',
+      save: 'Save',
+      saveChanges: 'Save changes',
+      createEnvironment: 'Create environment',
+      use: 'Use',
+      edit: 'Edit',
+      detect: 'Detect',
+      delete: 'Delete',
+      default: 'Default',
+      start: 'Start',
+      stop: 'Stop',
+      active: 'Active',
+      editing: 'Editing',
+      never: 'never',
+      notConfigured: 'Not configured',
+      notSelected: 'not selected',
+      unavailable: 'Unavailable',
+      available: 'Available',
+      writable: 'Writable',
+      notReady: 'Not ready',
+      configured: 'Configured',
+      detailLabel: 'Detail:',
+      noTags: 'No tags',
+      noDefaultWorkdir: 'No default workdir',
+      yes: 'yes',
+      no: 'no',
+      ok: '正常',
+      degraded: 'Degraded',
+      idle: 'Idle',
+      starting: 'Starting',
+      running: 'Running',
+      stopping: 'Stopping',
+      failed: 'Failed',
+      connecting: 'connecting',
+      connected: 'connected',
+      disconnected: 'disconnected',
+      error: 'error',
+      builtBy: 'Built by Kozmosa with ❤️',
+      collapse: 'Collapse',
+      confirmLogout: 'Confirm Logout',
+      confirmLogoutMessage: 'Are you sure you want to log out?',
+      dragToReorder: 'Drag to reorder',
+      expand: 'Expand',
+      logOut: 'Log out',
+      taskStatusLoading: 'Task | Loading…',
+      taskStatusSummary: 'Task | Total: {{total}}, Running: {{running}}, Pending: {{pending}}, Finished: {{finished}}',
+      taskStatusUnavailable: 'Task | Status unavailable',
+      loadingEditor: 'Loading editor...',
+      noMessages: 'No messages',
+    },
+    nav: {
+      literature: 'Literature',
+    },
+    literature: {
+      mySubscriptions: 'My Subscriptions',
+      newSubscription: 'New Subscription',
+      label: 'Label',
+      keywords: 'Keywords (comma separated)',
+      keywordsPlaceholder: 'machine learning, NLP',
+      categories: 'Categories',
+      frequency: 'Frequency',
+      daily: 'Daily',
+      weekly: 'Weekly',
+      twice_daily: 'Twice Daily',
+      create: 'Create',
+      delete: 'Delete',
+      unreadOnly: 'Unread Only',
+      refresh: 'Refresh',
+      markRead: 'Mark Read',
+      viewarXiv: 'arXiv',
+      convertToTask: 'Convert to Task',
+      convertConfirm: 'Convert this paper into an agent task. Select workspace, environment and execution engine below.',
+      convertSuccess: 'Paper converted to task successfully',
+      convertError: 'Failed to convert paper to task',
+      noPapers: 'No papers found. Create a subscription to get started.',
+      noSubscriptions: 'No subscriptions yet. Create one to track papers.',
+      aiPracticeNote: 'AI Practice Note',
+      aiSummary: 'AI Summary',
+    },
+};
+
+export default common;
