+++ frontend/src/shared/i18n/en/components.ts
@@ -1,0 +1,186 @@
+// components messages — English
+const components = {
+    components: {
+      healthStatusBar: {
+        loading: 'Loading system status...',
+        unable: 'Unable to fetch system status',
+        apiLabel: 'API:',
+        containerLabel: 'Container:',
+        sshLabel: 'SSH:',
+        workspaceLabel: 'Workspace:',
+      },
+      environmentSelector: {
+        eyebrow: 'Environment selector',
+        title: 'Shared environment selection',
+        description:
+          'Pick the execution environment once and reuse it across the frontend. Selection priority is: settings project default, remembered local choice, localhost seed, then the first available environment.',
+        selectedNone: 'No environment selected',
+        activeLabel: 'Active environment',
+        noAvailable: 'No environments available',
+        selectionStatus: 'Selection status',
+        loaded: 'Loaded:',
+        available: 'Available:',
+        loadError: 'Unable to load environments:',
+        displayName: 'Display name:',
+        host: 'Host:',
+        defaultWorkdir: 'Default workdir:',
+        waitingValidSelection: 'Waiting for a valid selection.',
+        createPrompt:
+          'Create an environment in Containers to enable terminal and workspace selection.',
+      },
+      terminalBench: {
+        eyebrow: '个人终端会话',
+        title: 'Personal terminal session',
+        description:
+          'Attach, detach, or reset the persistent personal terminal session for the selected environment, then open the interactive browser console.',
+        statusPrefix: 'Status:',
+        sessionSource: 'Session source',
+        loading: 'Loading:',
+        websocketUrl: 'WebSocket URL:',
+        environment: 'Environment:',
+        sessionName: 'Session name:',
+        attachment: 'Attachment:',
+        loadError: 'Load error:',
+        selectEnvironmentBeforeAttach: 'Select an environment before attaching the terminal session.',
+        noSessionYet: 'No personal terminal session has been created yet.',
+        detachedNotice: 'The personal session is still running, but this page is currently detached.',
+        attach: 'Attach',
+        detach: 'Detach',
+        resetSession: 'Reset session',
+        attaching: 'Attaching…',
+        detaching: 'Detaching…',
+        resetting: 'Resetting…',
+        consoleTitle: 'Terminal Console',
+      },
+      terminalConsole: {
+        startPrompt: 'Attach to a personal terminal session to open the xterm console.',
+        writeEnabled: 'Write enabled',
+        websocketSession: 'WebSocket session:',
+        attachment: 'Attachment:',
+        observeOnly: 'Observe-only',
+        connection: 'Connection:',
+        exited: 'Session exited with code {{code}}',
+      },
+      environmentEditor: {
+        eyebrow: 'Environment editor',
+        createTitle: 'Create environment',
+        editTitle: 'Edit environment',
+        description:
+          'Define SSH target details, runtime preferences, and the minimal metadata that powers the Containers control plane.',
+        activeLabel: 'Active environment:',
+        noActive: 'No environment has been marked as active yet.',
+        alias: 'Alias',
+        displayName: 'Display name',
+        descriptionField: 'Description',
+        host: 'Host',
+        port: 'Port',
+        user: 'User',
+        authKindLabel: 'Auth kind',
+        identityFile: 'Identity file',
+        proxyJump: 'Proxy jump',
+        proxyCommand: 'Proxy command',
+        tags: 'Tags',
+        defaultWorkdir: 'Default workdir',
+        preferredPython: 'Preferred Python',
+        preferredEnvManager: 'Preferred env manager',
+        preferredRuntimeNotes: 'Preferred runtime notes',
+        sshOptionsJson: 'SSH options JSON',
+        placeholders: {
+          alias: 'gpu-lab',
+          displayName: 'GPU Lab',
+          description: 'Primary training environment',
+          host: 'gpu.example.com',
+          port: '22',
+          user: 'root',
+          identityFile: '/keys/gpu-lab',
+          proxyJump: 'bastion',
+          proxyCommand: 'ssh -W %h:%p bastion',
+          tags: 'gpu, research',
+          defaultWorkdir: '/workspace/project',
+          preferredPython: 'python3.13',
+          preferredEnvManager: 'uv',
+          preferredRuntimeNotes: 'Use CUDA 12 image',
+          sshOptionsJson: '{"StrictHostKeyChecking":"no"}',
+        },
+        authKindOptions: {
+          ssh_key: 'SSH Key',
+          password: 'Password',
+          agent: 'Agent',
+        },
+        saveError: 'Unable to save environment',
+        portRangeError: 'Port must be between 1 and 65535',
+        noEnvironmentSelectedForEditing: 'No environment is selected for editing',
+        sshOptionsObjectError: 'SSH options must be a JSON object',
+        sshOptionsValuesError: 'SSH options values must be strings',
+        create: 'Create environment',
+        save: 'Save changes',
+        saving: 'Saving…',
+        reset: 'Reset',
+        cancelEdit: 'Cancel edit',
+      },
+      errorBoundary: {
+        title: 'Something went wrong',
+        retry: 'Try again',
+      },
+      modal: {
+        close: 'Close',
+      },
+      skills: {
+        ungrouped: 'Ungrouped',
+        mixed: 'Mixed',
+      },
+      token: {
+        tokens: 'Tokens',
+        total: 'Total:',
+        input: 'Input',
+        cache: 'Cache',
+        output: 'Output',
+        think: 'Think',
+      },
+      resources: {
+        cpu: 'CPU Usage',
+        cores: 'cores',
+        noGpu: 'No GPU detected',
+        memory: 'Memory',
+      },
+      environmentDetectionModal: {
+        groups: {
+          basicInfo: 'Basic Info',
+          pythonToolchain: 'Python Toolchain',
+          devTools: 'Dev Tools',
+          aiMl: 'AI / ML',
+          gpu: 'GPU Info',
+          envVars: 'Environment Variables',
+        },
+        labels: {
+          ssh: 'SSH',
+          tmux: 'Tmux',
+          hostname: 'Hostname',
+          os: 'OS',
+          arch: 'Architecture',
+          workdir: 'Workdir',
+          python: 'Python',
+          conda: 'Conda',
+          uv: 'uv',
+          pixi: 'pixi',
+          codex: 'Codex',
+          claudeCli: 'Claude CLI',
+          torch: 'PyTorch',
+          cuda: 'CUDA',
+          gpuCount: 'GPU Count',
+          gpuModels: 'GPU Models',
+          anthropic: 'Anthropic',
+        },
+        warningsHeader: 'Warnings',
+        status: {
+          available: 'Available',
+          unavailable: 'Unavailable',
+          present: 'Present',
+          missing: 'Missing',
+          unknown: 'Unknown',
+        },
+      },
+    },
+};
+
+export default components;
