+++ frontend/src/shared/i18n/en/index.ts
@@ -1,0 +1,5 @@
+export { default as common } from './common';
+export { default as auth } from './auth';
+export { default as navigation } from './navigation';
+export { default as pages } from './pages';
+export { default as components } from './components';
