+++ frontend/src/shared/i18n/en/navigation.ts
@@ -1,0 +1,59 @@
+// navigation messages — English
+const navigation = {
+    navigation: {
+      projects: {
+        label: 'Projects',
+        description: 'Project canvas and task graph visualization',
+      },
+      terminal: {
+        label: 'Terminal',
+        description: 'Health checks and terminal bench access',
+      },
+      tasks: {
+        label: 'Tasks',
+        description: 'Agent windows and observe-only terminals',
+      },
+      workspaces: {
+        label: 'Workspaces',
+        description: 'Project workspaces and task context',
+      },
+      workspaceBrowser: {
+        label: 'Browse Files',
+        description: 'Browse workspace files',
+      },
+      environments: {
+        label: 'Containers',
+        description: 'Runtime environments and container status',
+      },
+      resources: {
+        label: 'Resources',
+        description: 'System resource monitoring and usage',
+      },
+      sessions: {
+        label: 'Task Runs',
+        description: 'Task run history from real agent tasks',
+      },
+      timeline: {
+        label: 'Task Timeline',
+        description: 'Task run timeline from real agent tasks',
+      },
+      settings: {
+        label: 'Settings',
+        description: 'Runtime and WebUI preferences',
+      },
+    },
+    layout: {
+      brandLine:
+        'A unified control surface for research agents, environments, and workspaces.',
+      headerEyebrow: 'WebUI control surface',
+      headerDescription:
+        'Health, terminal bench, environment control plane, and workspace browser shell during backend realignment.',
+      footerDescription:
+        'Health, terminal bench, environment control plane, and workspace browser shell during backend realignment.',
+      collapseSidebar: 'Collapse sidebar',
+      expandSidebar: 'Expand sidebar',
+      resizeSidebar: 'Resize sidebar',
+    },
+};
+
+export default navigation;
