+++ frontend/src/shared/i18n/en/pages.ts
@@ -1,0 +1,721 @@
+// pages messages — English
+const pages = {
+    pages: {
+      projects: {
+        sidebarEyebrow: 'Projects',
+        sidebarTitle: 'Project Canvas',
+        sidebarCount: '{{count}} projects',
+        newProject: 'New project',
+        searchPlaceholder: 'Search projects',
+        newTask: 'New Task',
+        resetLayout: 'Reset Layout',
+        emptyCanvas: "Click 'New Task' to get started",
+        noProjects: 'No projects yet. Create one to start.',
+        moveToProject: 'Move to project',
+        noOtherProjects: 'No other projects to move to.',
+        createTitle: 'Create project',
+        createNameLabel: 'Project name',
+        createDescriptionLabel: 'Description (optional)',
+        createSubmit: 'Create project',
+      },
+      dashboard: {
+        eyebrow: 'Terminal',
+        title: 'Scholar Agent runtime shell',
+        description:
+          'The frontend now exposes the cleaned runtime shell: health checks, the keepalive personal terminal bench control surface, the environment control plane, and the managed workspace browser panel. Use this page to confirm API, SSH, workspace readiness, terminal session state, and environment readiness.',
+        surfaceTitle: 'Current backend surface',
+        endpointsLabel: 'Available endpoints:',
+        stateRootLabel: 'State root:',
+        detailLabel: 'Detail:',
+      },
+      terminal: {
+        eyebrow: '终端',
+        title: 'Terminal',
+        description:
+          'Attach to a persistent tmux session on your selected environment. Local environments connect directly; remote environments tunnel through interactive SSH.',
+        backToTasks: 'Back to tasks',
+        task: {
+          observeOnly: 'Observe-only',
+          takenOver: 'Taken over',
+          archived: 'Archived',
+          pending: 'Pending attach',
+          windowLabel: 'Window:',
+          ownerLabel: 'Owner:',
+          environmentLabel: 'Environment:',
+          noOwner: 'No active owner',
+          lastOutputLabel: 'Last output:',
+          otherOwner:
+            'This task is currently taken over by a different browser user. Release is unavailable here.',
+          placeholder: 'The task terminal attachment is not active yet.',
+        },
+      },
+      tasks: {
+        sidebarEyebrow: 'Task runs',
+        sidebarTitle: 'Agent tasks',
+        sidebarCount: '{{count}} total · live output replay',
+        newTask: 'New task',
+        resizeTaskList: 'Resize task list',
+        searchLabel: 'Search tasks',
+        searchPlaceholder: 'Search tasks',
+        empty: 'No tasks yet. Create one to start a run.',
+        noSearchResults: 'No tasks match “{{query}}”.',
+        updatedAt: 'Updated {{time}}',
+        createTitle: 'Create task',
+        createDescription:
+          'Launch a Claude Code task against a workspace and environment binding.',
+        closeCreate: 'Close create task',
+        workspaceLabel: 'Workspace',
+        environmentLabel: 'Environment',
+        projectLabel: 'Project',
+        profileLabel: 'Task profile',
+        researchAgentLabel: 'Research Agent',
+        taskConfigurationLabel: 'Task configuration',
+        skillsLabel: 'Skills',
+        skillDependenciesHint: 'Skills with dependencies will auto-include required skills at runtime.',
+        titleLabel: 'Title',
+        optionalPlaceholder: 'Optional',
+        taskInputLabel: 'Task input',
+        taskInputPlaceholder: 'Describe the task for the agent.',
+        structuredResearchGoal: 'Research goal',
+        structuredResearchContext: 'Context',
+        structuredResearchConstraints: 'Constraints',
+        structuredResearchDeliverables: 'Expected deliverables',
+        structuredResearchValidationPlan: 'Validation plan',
+        generatedPromptPreview: 'Generated prompt preview',
+        paperPathLabel: 'Paper path',
+        paperPathPlaceholder: 'papers/target.pdf',
+        scopeLabel: 'Scope',
+        targetTableLabel: 'Target table (optional)',
+        targetTablePlaceholder: 'Table3',
+        budgetHoursLabel: 'Budget (hours)',
+        researchTopicLabel: 'Research topic',
+        researchTopicPlaceholder: 'Graph neural networks for drug discovery',
+        seedPaperPathLabel: 'Seed paper path (optional)',
+        seedPaperPathPlaceholder: 'papers/seed.pdf',
+        depthLabel: 'Depth',
+        ideaSourceLabel: 'Idea source',
+        ideaSourcePlaceholder: 'Path to idea doc in workspace, or describe your idea here',
+        validationScopeLabel: 'Validation scope',
+        scopeCoreOnly: 'Core method',
+        scopeFullSuite: 'Full suite',
+        validationScopeQuick: 'Quick',
+        validationScopeFull: 'Full',
+        defaultWorkdir: 'Default workdir:',
+        resetDraft: 'Reset draft',
+        createAction: 'Create task',
+        creatingAction: 'Creating…',
+        create: {
+          taskPreset: 'Task preset',
+          presets: {
+            rawPrompt: 'Raw Prompt',
+            structuredResearch: 'Structured Research',
+            reproduceBaseline: 'Reproduce Baseline',
+            overview: 'Overview',
+          },
+          researcherType: 'Researcher Type',
+          researcherVanilla: 'Vanilla',
+          researcherAris: 'ARIS Researcher',
+          executionEngine: 'Execution Engine',
+          engineClaudeCode: 'Claude Code',
+          engineAgentSdk: 'Agent SDK',
+          engineCodexAppServer: 'Codex App Server',
+          skillsPlaceholder: 'analysis, code-review',
+          titlePlaceholder: 'Optional task title',
+          promptLabel: 'Prompt',
+          promptPlaceholder: 'Enter your research prompt...',
+          missingBinding: 'Select a workspace and environment before creating a task.',
+          noSkillsAvailable: 'No skills are available for this task.',
+          skillGroupSelected: '{{selected}}/{{total}} selected',
+          showSkillGroup: 'Show skills in {{group}}',
+          hideSkillGroup: 'Hide skills in {{group}}',
+          selectSkill: 'Select {{skill}}',
+          deselectSkill: 'Deselect {{skill}}',
+          seedFiles: {
+            dropzone: 'Drop seed papers here or click to browse (PDF, Markdown)',
+            uploading: 'Uploading {{name}}...',
+            uploadError: 'Failed to upload {{name}}: {{error}}',
+            removeFile: 'Remove {{name}}',
+            unsupportedType: 'Only PDF and Markdown files are accepted.',
+            maxFilesExceeded: 'Maximum {{max}} seed files allowed.',
+            fileCount: '{{count}} seed paper(s) uploaded',
+            label: 'Seed Papers',
+          },
+        },
+        noTaskSelected: 'No task selected',
+        noTaskSelectedDescription:
+          'Choose a task from the sidebar or create a new run to inspect its output.',
+        workspaceEyebrow: 'Task workspace',
+        outputTimeline: 'Output timeline',
+        latestSeq: 'Latest seq {{seq}}',
+        noOutput: 'No output has been recorded yet.',
+        promptLayers: 'Prompt layers',
+        prompt: 'Prompt',
+        chars: 'chars',
+        promptUnavailable: 'Prompt manifest is not available yet.',
+        summary: 'Summary',
+        taskId: 'Task ID',
+        created: 'Created',
+        updated: 'Updated',
+        started: 'Started',
+        completed: 'Completed',
+        binding: 'Binding',
+        profile: 'Profile',
+        workdir: 'Workdir',
+        snapshot: 'Snapshot',
+        bindingUnavailable: 'Binding snapshot is not available yet.',
+        runtime: 'Runtime',
+        runner: 'Runner',
+        directory: 'Directory',
+        command: 'Command',
+        runtimeUnavailable: 'Runtime payload is not available yet.',
+        result: 'Result',
+        exitCode: 'Exit code',
+        failure: 'Failure',
+        unavailable: 'n/a',
+        pendingValue: 'pending',
+        status: {
+          queued: 'Queued',
+          starting: 'Starting',
+          running: 'Running',
+          succeeded: 'Succeeded',
+          failed: 'Failed',
+          cancelled: 'Cancelled',
+          paused: 'Paused',
+        },
+        expandOutputTimeline: 'Expand output timeline',
+        collapseOutputTimeline: 'Collapse output timeline',
+        expandSummary: 'Expand summary',
+        collapseSummary: 'Collapse summary',
+        thinking: 'Thinking...',
+        toolCall: 'Tool: {{name}}',
+        toolResult: 'Result',
+        inputPlaceholder: 'Type a message...',
+        metadata: {
+          workdir: 'Workdir',
+          command: 'Command',
+          taskInput: 'Task input',
+        },
+        layout: {
+          showOnlyDetails: 'Show only details',
+          showOnlyConversation: 'Show only conversation',
+          collapseSidebar: 'Collapse details',
+          expandSidebar: 'Show details',
+          resizeSidebar: 'Resize details panel',
+          errorIndicator: 'Task has errors — click to view',
+        },
+        output: {
+          replayFailed: 'Unable to replay task output',
+          parseFailed: 'Unable to parse task output',
+          loadFailed: 'Unable to load task output',
+        },
+        messageGroup: {
+          summary: '{{count}} {{type}} messages',
+        },
+        messageType: {
+          systemEvent: 'system',
+          user: 'user',
+          assistant: 'assistant',
+          thinking: 'thinking',
+          toolCall: 'tool call',
+          toolResult: 'tool result',
+          stdout: 'stdout',
+          stderr: 'stderr',
+          lifecycle: 'lifecycle',
+        },
+        actions: {
+          cancel: 'Cancel',
+          archive: 'Archive',
+          retry: 'Retry',
+          showArchived: 'Show archived',
+          pause: 'Pause',
+          resume: 'Resume',
+          pauseFailed: 'Pause failed: {{error}}',
+          resumeFailed: 'Resume failed: {{error}}',
+          sendPromptFailed: 'Send prompt failed: {{error}}',
+          noTaskSelected: 'No task selected',
+          unexpectedError: 'An unexpected error occurred',
+        },
+        retrySuccess: 'Task retried successfully',
+        retryFailed: 'Failed to retry task',
+        sort: {
+          updated: 'Updated (newest)',
+          created: 'Created (newest)',
+          name: 'Name (A-Z)',
+        },
+      },
+      workspaces: {
+        eyebrow: '工作区',
+        title: 'Workspaces',
+        description:
+          'Manage reusable workspace contexts for task prompts, default workdirs, and research run boundaries.',
+        managerTitle: 'Workspace manager',
+        managerDescription:
+          'Create and maintain workspace records that tasks can bind to before launching in an environment.',
+        newWorkspace: 'New workspace',
+        createWorkspace: 'Create workspace',
+        saveWorkspace: 'Save workspace',
+        deleteWorkspace: 'Delete workspace',
+        confirmDelete: 'Confirm delete',
+        labelField: 'Workspace name',
+        descriptionField: 'Description',
+        defaultWorkdirField: 'Default workdir',
+        promptField: 'Workspace prompt',
+        noDefaultWorkdir: 'No default workdir',
+      },
+      workspaceBrowser: {
+        eyebrow: '工作区浏览器',
+        title: 'Workspace Browser',
+        description:
+          'Open the managed code-server browser for the environment selected in Settings.',
+        readinessTitle: 'Runtime setup required',
+        readinessMissingDependency: 'A runtime dependency is missing.',
+      },
+      settings: {
+        eyebrow: '设置',
+        title: 'Settings',
+        description:
+          'Customize your Scholar Agent experience: default landing page, terminal font size, project environment, and task draft templates.',
+        recoveryNotice:
+          'The local settings document was missing fields, invalid, or no longer compatible. Defaults were restored so the app can keep running.',
+        routes: {
+          terminal: 'Terminal',
+          tasks: 'Tasks',
+          workspaces: 'Workspaces',
+          environments: 'Environments',
+        },
+        general: {
+          title: 'General Preferences',
+          description:
+            'Preferences that apply across the entire interface.',
+          defaultRouteLabel: 'Default route',
+          terminalFontSizeLabel: 'Terminal font size',
+          terminalFontSizeHelp:
+            'Font size is clamped to a stable range of {{min}}-{{max}} px. Current save value: {{current}} px.',
+          editorFontSizeLabel: 'Editor font size',
+          editorFontFamilyLabel: 'Editor font family',
+          editorFontSizeHelp:
+            'Editor font size is clamped to {{min}}-{{max}} px. Current save value: {{current}} px.',
+        },
+        project: {
+          title: 'Project Defaults',
+          description:
+            'Defaults for the current project, stored locally in this browser.',
+          defaultEnvironmentLabel: 'Default environment',
+          defaultEnvironmentEmpty: 'No project default environment',
+          defaultEnvironmentHelp:
+            'This preferred environment is used before the remembered local selection.',
+          noEnvironments:
+            'Create at least one environment in Containers before configuring project defaults.',
+          environmentCardDescription:
+            'These fields only prefill new task drafts for this environment.',
+          titleTemplateLabel: 'Title template',
+          titleTemplatePlaceholder: 'Optional task title preset',
+          taskInputTemplateLabel: 'Task input template',
+          taskInputTemplatePlaceholder:
+            'Describe the default task input to preload for this environment.',
+          researchAgentDefaultLabel: 'Research Agent default',
+          taskConfigurationDefaultLabel: 'Task configuration default',
+        },
+        taskConfiguration: {
+          title: 'Task Configuration',
+          description:
+            'Configure the three task creation layers: Claude Code execution engine, Research Agent profile, and task input mode.',
+          executionEngineLabel: 'Execution engine',
+          defaultResearchAgentLabel: 'Default Research Agent',
+          defaultTaskConfigurationLabel: 'Default task configuration',
+          profileLabel: 'Research Agent label',
+          systemPromptLabel: 'System prompt',
+          skillsLabel: 'Enabled skills',
+          skillsDescription: 'Select the skills this research agent should use.',
+          skillsPromptLabel: 'Skills/config notes',
+          settingsJsonLabel: 'Claude settings JSON',
+          apiBaseUrlLabel: 'API Base URL',
+          apiKeyLabel: 'API Key',
+          defaultOpusModelLabel: 'Opus Model',
+          defaultSonnetModelLabel: 'Sonnet Model',
+          defaultHaikuModelLabel: 'Haiku Model',
+          envOverridesLabel: 'Env Overrides (JSON)',
+          placeholders: {
+            apiBaseUrl: 'https://api.anthropic.com/',
+            apiKey: 'sk-...',
+            opusModel: 'claude-opus-4-7',
+            sonnetModel: 'claude-sonnet-4-6',
+            haikuModel: 'claude-haiku-4-5',
+            envOverrides: '{\n  "CUSTOM_VAR": "value"\n}',
+          },
+        },
+        skillRepository: {
+          title: 'Skill Repository',
+          description: 'Manage AINRF skills: view details, preview settings, and import new skills.',
+          selectSkill: 'Select a skill to view details',
+          version: 'Version',
+          author: 'Author',
+          dependencies: 'Dependencies',
+          mcpServers: 'MCP Servers',
+          hooks: 'Hooks',
+          allowedAgents: 'Allowed Agents',
+          skillMdTitle: 'SKILL.md',
+          previewSettings: 'Preview Settings',
+          settingsPreviewTitle: 'Settings Preview',
+          importSkill: 'Import Skill',
+          importTitle: 'Import a Skill',
+          sourceLabel: 'Source',
+          gitSource: 'Git URL',
+          localSource: 'Local Path',
+          urlLabel: 'Git URL',
+          pathLabel: 'Local Path',
+          skillIdOverrideLabel: 'Skill ID Override (optional)',
+          importAction: 'Import',
+          importing: 'Importing…',
+          cancel: 'Cancel',
+          noSkills: 'No skills found in the repository.',
+          placeholders: {
+            url: 'https://github.com/example/skill-repo.git',
+            path: '/path/to/skill',
+          },
+        },
+        defaultWorkspace: {
+          title: 'Default Workspace',
+          description: 'Select the default workspace for task creation and file browsing.',
+          label: 'Default workspace',
+          noDefault: 'No default workspace',
+        },
+        searchBackend: {
+          title: 'Web Search Backend',
+          description: 'Choose how agent sessions access web search and content retrieval tools.',
+          activeBackend: 'Active backend',
+          autoStart: 'Auto-start MCP servers',
+          autoStartHint: 'Selected MCP servers will start automatically for every task session.',
+          native: 'Claude Native',
+          kindlyWebSearch: 'Kindly Web Search',
+          ccWebMcp: 'CC-Web-MCP',
+          saving: 'Saving…',
+          saved: 'Saved',
+        },
+        version: {
+          title: 'Deployment Versions',
+          description: 'Git commits baked into the running backend and frontend builds.',
+          backendLabel: 'Backend',
+          frontendLabel: 'Frontend',
+          commitLabel: 'Commit',
+          committedAtLabel: 'Committed at',
+          unavailable: 'Unavailable',
+          mismatchWarning: 'Backend and frontend were built from different commits.',
+        },
+        account: {
+          title: 'Account & Security',
+          description: 'Manage your account credentials.',
+        },
+        tabs: {
+          general: 'General',
+          llmProviders: 'LLM Providers',
+          monitoring: 'Monitoring & Audit',
+          users: 'User Management',
+          envAccess: 'Environment Access',
+          collaborators: 'Collaborators',
+        },
+        monitoring: {
+          title: 'Observability Platforms',
+          description: 'Quick access to monitoring, metrics, and LLM observability dashboards.',
+          noServices: 'No monitoring services are configured. Set the appropriate environment variables on the backend to enable links.',
+          notConfigured: 'Not configured',
+          openService: 'Open {{name}}',
+          grafanaDesc: 'Metrics dashboards, alerts, and visualization',
+          prometheusDesc: 'Time-series metrics collection and querying',
+          litefuseDesc: 'LLM observability: traces, generations, and token analytics',
+        },
+        envAccess: {
+          selectEnv: 'Select environment...',
+          maxTasks: 'Max tasks',
+          unlimited: 'unlimited',
+          grant: 'Grant',
+          remove: 'Remove',
+          grantTo: 'Grant to...',
+          noEnvSelected: 'Select an environment to manage access.',
+          noAccess: 'No access grants yet.',
+        },
+        collaborators: {
+          selectProject: 'Select project...',
+          add: 'Add',
+          addUser: 'Add user...',
+          remove: 'Remove',
+          role: {
+            member: 'Member',
+            viewer: 'Viewer',
+          },
+          noProjectSelected: 'Select a project to manage collaborators.',
+          noCollaborators: 'No collaborators yet.',
+          loadError: 'Failed to load collaborators',
+        },
+        engine: {
+          claudeCode: 'Claude Code',
+          claudeAgent: 'Agent SDK',
+          codexAppServer: 'Codex App Server',
+        },
+        codex: {
+          model: 'Model',
+          command: 'Command',
+          approval: 'Approval',
+          baseUrl: 'Base URL',
+          apiKey: 'API Key',
+          config: 'Codex config.toml',
+          auth: 'Codex auth.json',
+          install: 'Install {{name}}',
+          installing: 'Installing {{name}}…',
+          update: 'Update {{name}}',
+          updating: 'Updating…',
+          installed: '{{name}} installed',
+          updateTitle: 'Update {{name}}',
+          updateWarning: 'This will replace existing files. Continue?',
+          forceUpdate: 'Force Update',
+          placeholders: {
+            baseUrl: 'https://api.openai.com/',
+            apiKey: 'sk-...',
+            model: 'gpt-5-codex',
+            command: 'codex app-server --listen stdio://',
+            approval: 'never',
+            config: 'model = "gpt-5-codex"',
+            auth: '{\n  "OPENAI_API_KEY": "sk-..."\n}',
+          },
+        },
+        appearance: {
+          title: 'Appearance',
+          description: 'Global font family preference',
+          fontFamilyLabel: 'Font family',
+          sansSerif: 'Sans-serif',
+          serif: 'Serif',
+          previewHint: 'Changes apply immediately across the entire interface.',
+        },
+        users: {
+          title: 'User Management',
+          count: '{{count}} users',
+          approve: 'Approve',
+          disable: 'Disable',
+          reEnable: 'Re-enable',
+          resetPassword: 'Reset Password',
+          newPassword: 'New Password:',
+          enterNewPassword: 'Enter new password',
+          set: 'Set',
+        },
+        llmProviders: {
+          title: 'LLM Providers',
+          description: 'Manage global LLM API configurations for quick profile fill.',
+          addProvider: 'Add Provider',
+          editProvider: 'Edit Provider',
+          deleteProvider: 'Delete Provider',
+          noProviders: 'No providers configured yet. Add one to enable quick fill in agent profiles.',
+          nameLabel: 'Name',
+          formatLabel: 'Format',
+          baseUrlLabel: 'Base URL',
+          apiKeyLabel: 'API Key',
+          opusModelLabel: 'Opus Model',
+          sonnetModelLabel: 'Sonnet Model',
+          haikuModelLabel: 'Haiku Model',
+          defaultModelLabel: 'Default Model',
+          formatAnthropic: 'Anthropic',
+          formatOpenAIChat: 'OpenAI Chat',
+          formatOpenAIResponses: 'OpenAI Responses',
+          confirmDelete: 'Delete provider "{{name}}"?',
+          fillFromProvider: 'Fill from provider',
+          customOption: 'Custom',
+          placeholders: {
+            name: 'e.g. Kimi Coding',
+            anthropicBaseUrl: 'https://api.anthropic.com/',
+            openAiBaseUrl: 'https://api.openai.com/',
+            apiKey: 'sk-...',
+            opusModel: 'claude-opus-4-7',
+            sonnetModel: 'claude-sonnet-4-6',
+            haikuModel: 'claude-haiku-4-5',
+            defaultModel: 'gpt-4o',
+          },
+          updateFailed: 'Update failed',
+        },
+      },
+      placeholder: {
+        eyebrow: 'Settings',
+        title: 'Settings are coming soon',
+        description:
+          'This area will host environment preferences, runtime configuration, and WebUI behavior controls in a later slice.',
+        badge: 'Placeholder section',
+      },
+      environments: {
+        eyebrow: '容器',
+        title: 'Containers',
+        description:
+          'Add and manage SSH-backed compute environments. Run detection to verify connectivity, then mark one as active for terminals and tasks.',
+        currentSelection: 'Current selection',
+        activeSelectionFallback: 'No active environment selected yet',
+        addEnvironment: 'Add environment',
+        listTitle: 'Environment list',
+        listDescription:
+          'Use the list to inspect the current control-plane state, run detection, or mark an environment active for downstream pages.',
+        loading: 'Loading environments...',
+        empty: 'No environments have been created yet.',
+        alias: 'Alias',
+        host: 'Host',
+        auth: 'Auth',
+        detection: 'Detection',
+        actions: 'Actions',
+        defaultWorkdir: 'No default workdir',
+        activeBadge: 'Active',
+        editingBadge: 'Editing',
+        detectedAt: 'Detected',
+        notDetected: 'Not detected yet',
+        confirmDelete: 'Delete environment "{{alias}}"?',
+        defaultEnvironmentLocked: 'The default localhost environment cannot be deleted.',
+        detectFallbackToast: '{{alias}} detection fell back to personal tmux after SSH was unavailable.',
+        detectFailedToast: 'Environment detection failed.',
+        detectionStatus: {
+          success: 'Success',
+          partial: 'Partial',
+          failed: 'Failed',
+        },
+        authKind: {
+          ssh_key: 'SSH key',
+          password: 'Password',
+          agent: 'Agent',
+        },
+        projectReferenceTitle: 'Project environment binding',
+        projectReferenceDescription:
+          'Attach the selected environment to the current project, mark it as the project default, and override runtime-only fields such as workdir or env manager.',
+        projectReferenceNoSelection: 'Select an environment from the list before editing project bindings.',
+        projectReferencedBadge: 'Project ref',
+        projectUnreferencedBadge: 'Not referenced',
+        projectDefaultBadge: 'Project default',
+        setProjectDefault: 'Default',
+        setProjectDefaultDisabled: 'Is default',
+        useTooltip: 'Use as session default',
+        defaultTooltip: 'Set as new project default',
+        clearProjectDefault: 'Clear project default',
+        removeProjectReference: 'Remove project ref',
+        projectReferenceSaveError: 'Unable to save project environment binding',
+        projectOverrideWorkdir: 'Override workdir',
+        projectOverrideWorkdirPlaceholder: '/workspace/project-override',
+        projectOverrideEnvName: 'Override env name',
+        projectOverrideEnvNamePlaceholder: 'research-env',
+        projectOverrideEnvManager: 'Override env manager',
+        projectOverrideEnvManagerPlaceholder: 'conda',
+        projectOverrideRuntimeNotes: 'Override runtime notes',
+        projectOverrideRuntimeNotesPlaceholder: 'Prefer the project-specific runtime image.',
+        projectReferenceSaving: 'Saving binding…',
+        updateProjectReference: 'Update project binding',
+        attachProjectReference: 'Attach to project',
+      },
+      resources: {
+        eyebrow: 'RESOURCES',
+        title: 'Resource Monitor',
+        description: 'Real-time system resource usage across all environments.',
+        loading: 'Loading resources...',
+        noData: 'No resource data available yet.',
+        systemCard: {
+          title: 'System Resources',
+          gpuTitle: 'GPU',
+          cpuTitle: 'CPU',
+          memoryTitle: 'Memory',
+          noGpu: 'No GPU detected',
+        },
+        processCard: {
+          title: 'AINRF Processes',
+          empty: 'No AINRF processes detected',
+          columns: {
+            pid: 'PID',
+            name: 'Name',
+            cpu: 'CPU%',
+            memory: 'Memory',
+            runtime: 'Runtime',
+          },
+        },
+        taskUsage: {
+          title: 'Task Usage',
+          description: 'Token and duration totals across all tasks.',
+          taskCount: '{{count}} tasks',
+          totalTokens: 'Total tokens',
+          totalDuration: 'Total task duration',
+          medianDuration: 'Median task duration',
+          topTasks: 'Top token tasks',
+          tokens: '{{count}} tokens',
+          noTopTasks: 'No task token usage recorded yet.',
+        },
+        status: {
+          ok: 'Healthy',
+          degraded: 'Degraded',
+          unavailable: 'Unavailable',
+        },
+      },
+      sessions: {
+        sidebarTitle: 'Task runs',
+        sidebarCount: '{{count}} runs',
+        searchPlaceholder: 'Search task runs...',
+        empty: 'No task runs yet',
+        selectPrompt: 'Select a task run to view details',
+        notFound: 'Task run not found',
+        outputCount: '{{count}} output events',
+        taskCount: '{{count}} tasks',
+        attemptsTitle: 'Attempts',
+        noAttempts: 'No attempts recorded',
+        attemptLabel: 'Attempt #{{seq}}',
+        viewTask: 'View task',
+        hasTokens: 'Token data',
+        status: {
+          active: 'Active',
+          completed: 'Completed',
+          archived: 'Archived',
+        },
+        attemptStatus: {
+          running: 'Running',
+          completed: 'Completed',
+          failed: 'Failed',
+          interrupted: 'Interrupted',
+        },
+        perModelBreakdown: 'Per-Model Breakdown',
+        detail: {
+          prompt: 'Prompt',
+          engine: 'Engine',
+          project: 'Project',
+          workspace: 'Workspace',
+          environment: 'Environment',
+          workdir: 'Workdir',
+          command: 'Command',
+          started: 'Started',
+          completed: 'Completed',
+          outputSeq: 'Output events',
+          exitCode: 'Exit code',
+        },
+        totalTokens: '{{count}} tokens',
+        fileBrowser: {
+          emptyDirectory: 'Empty directory',
+          loading: 'Loading files...',
+          files: 'Files',
+          noFileSelected: 'No file selected',
+          noFiles: 'No files',
+          loadingFile: 'Loading file...',
+          selectEnv: 'Select an environment to browse files',
+          refresh: 'Refresh',
+          selectFile: 'Select a file to view its contents',
+          binaryFile: 'Binary file',
+        },
+        timeline: {
+          title: 'Task run timeline',
+          empty: 'No task runs in this time range',
+          loading: 'Loading task timeline...',
+          taskCount: '{{count}} task runs',
+          projectFilter: 'Filter task timeline by project',
+          allProjects: 'All Projects',
+          today: 'Today',
+          past7Days: 'Past 7 days',
+          past30Days: 'Past 30 days',
+          from: 'From',
+          to: 'To',
+          attemptSeq: 'Attempt #{seq}',
+          status: 'Status:',
+          duration: 'Duration:',
+          reason: 'Reason:',
+          attemptLabel: 'Attempt #{{seq}}',
+          viewTask: 'View task',
+          viewSession: 'View session',
+        },
+      },
+    },
+};
+
+export default pages;
