+++ frontend/src/shared/i18n/messages.ts
@@ -16,2144 +16,7 @@ type Paths<T, Prefix extends string = ''> = T extends string
           : never;
     }[Extract<keyof T, string>];
 
-export const messages = {
-  en: {
-    chat: {
-      thinking: 'Thinking process',
-      disclaimer: 'AI can make mistakes. Please verify important information.',
-      inputPlaceholder: 'Message AI...',
-      send: 'Send message',
-      attach: 'Attach file',
-      webSearch: 'Web search',
-      reason: 'Reason',
-      copy: 'Copy',
-      retry: 'Retry',
-      collapseTools: 'Collapse tools',
-      scrollToBottom: 'Scroll to bottom',
-    },
-    common: {
-      appName: 'AINRF Console',
-      language: 'Language',
-      english: 'English',
-      chinese: '中文',
-      loading: 'Loading...',
-      saving: 'Saving…',
-      reset: 'Reset',
-      cancel: 'Cancel',
-      save: 'Save',
-      saveChanges: 'Save changes',
-      createEnvironment: 'Create environment',
-      use: 'Use',
-      edit: 'Edit',
-      detect: 'Detect',
-      delete: 'Delete',
-      default: 'Default',
-      start: 'Start',
-      stop: 'Stop',
-      active: 'Active',
-      editing: 'Editing',
-      never: 'never',
-      notConfigured: 'Not configured',
-      notSelected: 'not selected',
-      unavailable: 'Unavailable',
-      available: 'Available',
-      writable: 'Writable',
-      notReady: 'Not ready',
-      configured: 'Configured',
-      detailLabel: 'Detail:',
-      noTags: 'No tags',
-      noDefaultWorkdir: 'No default workdir',
-      yes: 'yes',
-      no: 'no',
-      ok: '正常',
-      degraded: 'Degraded',
-      idle: 'Idle',
-      starting: 'Starting',
-      running: 'Running',
-      stopping: 'Stopping',
-      failed: 'Failed',
-      connecting: 'connecting',
-      connected: 'connected',
-      disconnected: 'disconnected',
-      error: 'error',
-      builtBy: 'Built by Kozmosa with ❤️',
-      collapse: 'Collapse',
-      confirmLogout: 'Confirm Logout',
-      confirmLogoutMessage: 'Are you sure you want to log out?',
-      dragToReorder: 'Drag to reorder',
-      expand: 'Expand',
-      logOut: 'Log out',
-      taskStatusLoading: 'Task | Loading…',
-      taskStatusSummary: 'Task | Total: {{total}}, Running: {{running}}, Pending: {{pending}}, Finished: {{finished}}',
-      taskStatusUnavailable: 'Task | Status unavailable',
-      loadingEditor: 'Loading editor...',
-      noMessages: 'No messages',
-    },
-    nav: {
-      literature: 'Literature',
-    },
-    literature: {
-      mySubscriptions: 'My Subscriptions',
-      newSubscription: 'New Subscription',
-      label: 'Label',
-      keywords: 'Keywords (comma separated)',
-      keywordsPlaceholder: 'machine learning, NLP',
-      categories: 'Categories',
-      frequency: 'Frequency',
-      daily: 'Daily',
-      weekly: 'Weekly',
-      twice_daily: 'Twice Daily',
-      create: 'Create',
-      delete: 'Delete',
-      unreadOnly: 'Unread Only',
-      refresh: 'Refresh',
-      markRead: 'Mark Read',
-      viewarXiv: 'arXiv',
-      convertToTask: 'Convert to Task',
-      convertConfirm: 'Convert this paper into an agent task. Select workspace, environment and execution engine below.',
-      convertSuccess: 'Paper converted to task successfully',
-      convertError: 'Failed to convert paper to task',
-      noPapers: 'No papers found. Create a subscription to get started.',
-      noSubscriptions: 'No subscriptions yet. Create one to track papers.',
-      aiPracticeNote: 'AI Practice Note',
-      aiSummary: 'AI Summary',
-    },
-    auth: {
-      login: 'Log in',
-      logout: 'Log out',
-      username: 'Username',
-      usernameHint: 'letters, digits, dot, underscore, hyphen only',
-      password: 'Password',
-      displayName: 'Display Name',
-      confirmPassword: 'Confirm Password',
-      passwordMismatch: 'Passwords do not match',
-      registerFailed: 'Registration failed',
-      registrationSubmitted: 'Registration Submitted',
-      backToLogin: 'Back to Login',
-      registerLink: "Don't have an account? Register",
-      loginLink: 'Already have an account? Log in',
-      register: 'Register',
-      pendingApproval: 'Your account is pending admin approval.',
-      changePassword: 'Change Password',
-      mustChangePassword: 'You must change your password before continuing.',
-      currentPassword: 'Current Password',
-      newPassword: 'New Password',
-      changePasswordFailed: 'Failed to change password',
-      passwordTooShort: 'Password must be at least 4 characters',
-      loginFailed: 'Login failed',
-    },
-    navigation: {
-      projects: {
-        label: 'Projects',
-        description: 'Project canvas and task graph visualization',
-      },
-      terminal: {
-        label: 'Terminal',
-        description: 'Health checks and terminal bench access',
-      },
-      tasks: {
-        label: 'Tasks',
-        description: 'Agent windows and observe-only terminals',
-      },
-      workspaces: {
-        label: 'Workspaces',
-        description: 'Project workspaces and task context',
-      },
-      workspaceBrowser: {
-        label: 'Browse Files',
-        description: 'Browse workspace files',
-      },
-      environments: {
-        label: 'Containers',
-        description: 'Runtime environments and container status',
-      },
-      resources: {
-        label: 'Resources',
-        description: 'System resource monitoring and usage',
-      },
-      sessions: {
-        label: 'Task Runs',
-        description: 'Task run history from real agent tasks',
-      },
-      timeline: {
-        label: 'Task Timeline',
-        description: 'Task run timeline from real agent tasks',
-      },
-      settings: {
-        label: 'Settings',
-        description: 'Runtime and WebUI preferences',
-      },
-    },
-    layout: {
-      brandLine:
-        'A unified control surface for research agents, environments, and workspaces.',
-      headerEyebrow: 'WebUI control surface',
-      headerDescription:
-        'Health, terminal bench, environment control plane, and workspace browser shell during backend realignment.',
-      footerDescription:
-        'Health, terminal bench, environment control plane, and workspace browser shell during backend realignment.',
-      collapseSidebar: 'Collapse sidebar',
-      expandSidebar: 'Expand sidebar',
-      resizeSidebar: 'Resize sidebar',
-    },
-    pages: {
-      projects: {
-        sidebarEyebrow: 'Projects',
-        sidebarTitle: 'Project Canvas',
-        sidebarCount: '{{count}} projects',
-        newProject: 'New project',
-        searchPlaceholder: 'Search projects',
-        newTask: 'New Task',
-        resetLayout: 'Reset Layout',
-        emptyCanvas: "Click 'New Task' to get started",
-        noProjects: 'No projects yet. Create one to start.',
-        moveToProject: 'Move to project',
-        noOtherProjects: 'No other projects to move to.',
-        createTitle: 'Create project',
-        createNameLabel: 'Project name',
-        createDescriptionLabel: 'Description (optional)',
-        createSubmit: 'Create project',
-      },
-      dashboard: {
-        eyebrow: 'Terminal',
-        title: 'Scholar Agent runtime shell',
-        description:
-          'The frontend now exposes the cleaned runtime shell: health checks, the keepalive personal terminal bench control surface, the environment control plane, and the managed workspace browser panel. Use this page to confirm API, SSH, workspace readiness, terminal session state, and environment readiness.',
-        surfaceTitle: 'Current backend surface',
-        endpointsLabel: 'Available endpoints:',
-        stateRootLabel: 'State root:',
-        detailLabel: 'Detail:',
-      },
-      terminal: {
-        eyebrow: '终端',
-        title: 'Terminal',
-        description:
-          'Attach to a persistent tmux session on your selected environment. Local environments connect directly; remote environments tunnel through interactive SSH.',
-        backToTasks: 'Back to tasks',
-        task: {
-          observeOnly: 'Observe-only',
-          takenOver: 'Taken over',
-          archived: 'Archived',
-          pending: 'Pending attach',
-          windowLabel: 'Window:',
-          ownerLabel: 'Owner:',
-          environmentLabel: 'Environment:',
-          noOwner: 'No active owner',
-          lastOutputLabel: 'Last output:',
-          otherOwner:
-            'This task is currently taken over by a different browser user. Release is unavailable here.',
-          placeholder: 'The task terminal attachment is not active yet.',
-        },
-      },
-      tasks: {
-        sidebarEyebrow: 'Task runs',
-        sidebarTitle: 'Agent tasks',
-        sidebarCount: '{{count}} total · live output replay',
-        newTask: 'New task',
-        resizeTaskList: 'Resize task list',
-        searchLabel: 'Search tasks',
-        searchPlaceholder: 'Search tasks',
-        empty: 'No tasks yet. Create one to start a run.',
-        noSearchResults: 'No tasks match “{{query}}”.',
-        updatedAt: 'Updated {{time}}',
-        createTitle: 'Create task',
-        createDescription:
-          'Launch a Claude Code task against a workspace and environment binding.',
-        closeCreate: 'Close create task',
-        workspaceLabel: 'Workspace',
-        environmentLabel: 'Environment',
-        projectLabel: 'Project',
-        profileLabel: 'Task profile',
-        researchAgentLabel: 'Research Agent',
-        taskConfigurationLabel: 'Task configuration',
-        skillsLabel: 'Skills',
-        skillDependenciesHint: 'Skills with dependencies will auto-include required skills at runtime.',
-        titleLabel: 'Title',
-        optionalPlaceholder: 'Optional',
-        taskInputLabel: 'Task input',
-        taskInputPlaceholder: 'Describe the task for the agent.',
-        structuredResearchGoal: 'Research goal',
-        structuredResearchContext: 'Context',
-        structuredResearchConstraints: 'Constraints',
-        structuredResearchDeliverables: 'Expected deliverables',
-        structuredResearchValidationPlan: 'Validation plan',
-        generatedPromptPreview: 'Generated prompt preview',
-        paperPathLabel: 'Paper path',
-        paperPathPlaceholder: 'papers/target.pdf',
-        scopeLabel: 'Scope',
-        targetTableLabel: 'Target table (optional)',
-        targetTablePlaceholder: 'Table3',
-        budgetHoursLabel: 'Budget (hours)',
-        researchTopicLabel: 'Research topic',
-        researchTopicPlaceholder: 'Graph neural networks for drug discovery',
-        seedPaperPathLabel: 'Seed paper path (optional)',
-        seedPaperPathPlaceholder: 'papers/seed.pdf',
-        depthLabel: 'Depth',
-        ideaSourceLabel: 'Idea source',
-        ideaSourcePlaceholder: 'Path to idea doc in workspace, or describe your idea here',
-        validationScopeLabel: 'Validation scope',
-        scopeCoreOnly: 'Core method',
-        scopeFullSuite: 'Full suite',
-        validationScopeQuick: 'Quick',
-        validationScopeFull: 'Full',
-        defaultWorkdir: 'Default workdir:',
-        resetDraft: 'Reset draft',
-        createAction: 'Create task',
-        creatingAction: 'Creating…',
-        create: {
-          taskPreset: 'Task preset',
-          presets: {
-            rawPrompt: 'Raw Prompt',
-            structuredResearch: 'Structured Research',
-            reproduceBaseline: 'Reproduce Baseline',
-            overview: 'Overview',
-          },
-          researcherType: 'Researcher Type',
-          researcherVanilla: 'Vanilla',
-          researcherAris: 'ARIS Researcher',
-          executionEngine: 'Execution Engine',
-          engineClaudeCode: 'Claude Code',
-          engineAgentSdk: 'Agent SDK',
-          engineCodexAppServer: 'Codex App Server',
-          skillsPlaceholder: 'analysis, code-review',
-          titlePlaceholder: 'Optional task title',
-          promptLabel: 'Prompt',
-          promptPlaceholder: 'Enter your research prompt...',
-          missingBinding: 'Select a workspace and environment before creating a task.',
-          noSkillsAvailable: 'No skills are available for this task.',
-          skillGroupSelected: '{{selected}}/{{total}} selected',
-          showSkillGroup: 'Show skills in {{group}}',
-          hideSkillGroup: 'Hide skills in {{group}}',
-          selectSkill: 'Select {{skill}}',
-          deselectSkill: 'Deselect {{skill}}',
-          seedFiles: {
-            dropzone: 'Drop seed papers here or click to browse (PDF, Markdown)',
-            uploading: 'Uploading {{name}}...',
-            uploadError: 'Failed to upload {{name}}: {{error}}',
-            removeFile: 'Remove {{name}}',
-            unsupportedType: 'Only PDF and Markdown files are accepted.',
-            maxFilesExceeded: 'Maximum {{max}} seed files allowed.',
-            fileCount: '{{count}} seed paper(s) uploaded',
-            label: 'Seed Papers',
-          },
-        },
-        noTaskSelected: 'No task selected',
-        noTaskSelectedDescription:
-          'Choose a task from the sidebar or create a new run to inspect its output.',
-        workspaceEyebrow: 'Task workspace',
-        outputTimeline: 'Output timeline',
-        latestSeq: 'Latest seq {{seq}}',
-        noOutput: 'No output has been recorded yet.',
-        promptLayers: 'Prompt layers',
-        prompt: 'Prompt',
-        chars: 'chars',
-        promptUnavailable: 'Prompt manifest is not available yet.',
-        summary: 'Summary',
-        taskId: 'Task ID',
-        created: 'Created',
-        updated: 'Updated',
-        started: 'Started',
-        completed: 'Completed',
-        binding: 'Binding',
-        profile: 'Profile',
-        workdir: 'Workdir',
-        snapshot: 'Snapshot',
-        bindingUnavailable: 'Binding snapshot is not available yet.',
-        runtime: 'Runtime',
-        runner: 'Runner',
-        directory: 'Directory',
-        command: 'Command',
-        runtimeUnavailable: 'Runtime payload is not available yet.',
-        result: 'Result',
-        exitCode: 'Exit code',
-        failure: 'Failure',
-        unavailable: 'n/a',
-        pendingValue: 'pending',
-        status: {
-          queued: 'Queued',
-          starting: 'Starting',
-          running: 'Running',
-          succeeded: 'Succeeded',
-          failed: 'Failed',
-          cancelled: 'Cancelled',
-          paused: 'Paused',
-        },
-        expandOutputTimeline: 'Expand output timeline',
-        collapseOutputTimeline: 'Collapse output timeline',
-        expandSummary: 'Expand summary',
-        collapseSummary: 'Collapse summary',
-        thinking: 'Thinking...',
-        toolCall: 'Tool: {{name}}',
-        toolResult: 'Result',
-        inputPlaceholder: 'Type a message...',
-        metadata: {
-          workdir: 'Workdir',
-          command: 'Command',
-          taskInput: 'Task input',
-        },
-        layout: {
-          showOnlyDetails: 'Show only details',
-          showOnlyConversation: 'Show only conversation',
-          collapseSidebar: 'Collapse details',
-          expandSidebar: 'Show details',
-          resizeSidebar: 'Resize details panel',
-          errorIndicator: 'Task has errors — click to view',
-        },
-        output: {
-          replayFailed: 'Unable to replay task output',
-          parseFailed: 'Unable to parse task output',
-          loadFailed: 'Unable to load task output',
-        },
-        messageGroup: {
-          summary: '{{count}} {{type}} messages',
-        },
-        messageType: {
-          systemEvent: 'system',
-          user: 'user',
-          assistant: 'assistant',
-          thinking: 'thinking',
-          toolCall: 'tool call',
-          toolResult: 'tool result',
-          stdout: 'stdout',
-          stderr: 'stderr',
-          lifecycle: 'lifecycle',
-        },
-        actions: {
-          cancel: 'Cancel',
-          archive: 'Archive',
-          retry: 'Retry',
-          showArchived: 'Show archived',
-          pause: 'Pause',
-          resume: 'Resume',
-          pauseFailed: 'Pause failed: {{error}}',
-          resumeFailed: 'Resume failed: {{error}}',
-          sendPromptFailed: 'Send prompt failed: {{error}}',
-          noTaskSelected: 'No task selected',
-          unexpectedError: 'An unexpected error occurred',
-        },
-        retrySuccess: 'Task retried successfully',
-        retryFailed: 'Failed to retry task',
-        sort: {
-          updated: 'Updated (newest)',
-          created: 'Created (newest)',
-          name: 'Name (A-Z)',
-        },
-      },
-      workspaces: {
-        eyebrow: '工作区',
-        title: 'Workspaces',
-        description:
-          'Manage reusable workspace contexts for task prompts, default workdirs, and research run boundaries.',
-        managerTitle: 'Workspace manager',
-        managerDescription:
-          'Create and maintain workspace records that tasks can bind to before launching in an environment.',
-        newWorkspace: 'New workspace',
-        createWorkspace: 'Create workspace',
-        saveWorkspace: 'Save workspace',
-        deleteWorkspace: 'Delete workspace',
-        confirmDelete: 'Confirm delete',
-        labelField: 'Workspace name',
-        descriptionField: 'Description',
-        defaultWorkdirField: 'Default workdir',
-        promptField: 'Workspace prompt',
-        noDefaultWorkdir: 'No default workdir',
-      },
-      workspaceBrowser: {
-        eyebrow: '工作区浏览器',
-        title: 'Workspace Browser',
-        description:
-          'Open the managed code-server browser for the environment selected in Settings.',
-        readinessTitle: 'Runtime setup required',
-        readinessMissingDependency: 'A runtime dependency is missing.',
-      },
-      settings: {
-        eyebrow: '设置',
-        title: 'Settings',
-        description:
-          'Customize your Scholar Agent experience: default landing page, terminal font size, project environment, and task draft templates.',
-        recoveryNotice:
-          'The local settings document was missing fields, invalid, or no longer compatible. Defaults were restored so the app can keep running.',
-        routes: {
-          terminal: 'Terminal',
-          tasks: 'Tasks',
-          workspaces: 'Workspaces',
-          environments: 'Environments',
-        },
-        general: {
-          title: 'General Preferences',
-          description:
-            'Preferences that apply across the entire interface.',
-          defaultRouteLabel: 'Default route',
-          terminalFontSizeLabel: 'Terminal font size',
-          terminalFontSizeHelp:
-            'Font size is clamped to a stable range of {{min}}-{{max}} px. Current save value: {{current}} px.',
-          editorFontSizeLabel: 'Editor font size',
-          editorFontFamilyLabel: 'Editor font family',
-          editorFontSizeHelp:
-            'Editor font size is clamped to {{min}}-{{max}} px. Current save value: {{current}} px.',
-        },
-        project: {
-          title: 'Project Defaults',
-          description:
-            'Defaults for the current project, stored locally in this browser.',
-          defaultEnvironmentLabel: 'Default environment',
-          defaultEnvironmentEmpty: 'No project default environment',
-          defaultEnvironmentHelp:
-            'This preferred environment is used before the remembered local selection.',
-          noEnvironments:
-            'Create at least one environment in Containers before configuring project defaults.',
-          environmentCardDescription:
-            'These fields only prefill new task drafts for this environment.',
-          titleTemplateLabel: 'Title template',
-          titleTemplatePlaceholder: 'Optional task title preset',
-          taskInputTemplateLabel: 'Task input template',
-          taskInputTemplatePlaceholder:
-            'Describe the default task input to preload for this environment.',
-          researchAgentDefaultLabel: 'Research Agent default',
-          taskConfigurationDefaultLabel: 'Task configuration default',
-        },
-        taskConfiguration: {
-          title: 'Task Configuration',
-          description:
-            'Configure the three task creation layers: Claude Code execution engine, Research Agent profile, and task input mode.',
-          executionEngineLabel: 'Execution engine',
-          defaultResearchAgentLabel: 'Default Research Agent',
-          defaultTaskConfigurationLabel: 'Default task configuration',
-          profileLabel: 'Research Agent label',
-          systemPromptLabel: 'System prompt',
-          skillsLabel: 'Enabled skills',
-          skillsDescription: 'Select the skills this research agent should use.',
-          skillsPromptLabel: 'Skills/config notes',
-          settingsJsonLabel: 'Claude settings JSON',
-          apiBaseUrlLabel: 'API Base URL',
-          apiKeyLabel: 'API Key',
-          defaultOpusModelLabel: 'Opus Model',
-          defaultSonnetModelLabel: 'Sonnet Model',
-          defaultHaikuModelLabel: 'Haiku Model',
-          envOverridesLabel: 'Env Overrides (JSON)',
-          placeholders: {
-            apiBaseUrl: 'https://api.anthropic.com/',
-            apiKey: 'sk-...',
-            opusModel: 'claude-opus-4-7',
-            sonnetModel: 'claude-sonnet-4-6',
-            haikuModel: 'claude-haiku-4-5',
-            envOverrides: '{\n  "CUSTOM_VAR": "value"\n}',
-          },
-        },
-        skillRepository: {
-          title: 'Skill Repository',
-          description: 'Manage AINRF skills: view details, preview settings, and import new skills.',
-          selectSkill: 'Select a skill to view details',
-          version: 'Version',
-          author: 'Author',
-          dependencies: 'Dependencies',
-          mcpServers: 'MCP Servers',
-          hooks: 'Hooks',
-          allowedAgents: 'Allowed Agents',
-          skillMdTitle: 'SKILL.md',
-          previewSettings: 'Preview Settings',
-          settingsPreviewTitle: 'Settings Preview',
-          importSkill: 'Import Skill',
-          importTitle: 'Import a Skill',
-          sourceLabel: 'Source',
-          gitSource: 'Git URL',
-          localSource: 'Local Path',
-          urlLabel: 'Git URL',
-          pathLabel: 'Local Path',
-          skillIdOverrideLabel: 'Skill ID Override (optional)',
-          importAction: 'Import',
-          importing: 'Importing…',
-          cancel: 'Cancel',
-          noSkills: 'No skills found in the repository.',
-          placeholders: {
-            url: 'https://github.com/example/skill-repo.git',
-            path: '/path/to/skill',
-          },
-        },
-        defaultWorkspace: {
-          title: 'Default Workspace',
-          description: 'Select the default workspace for task creation and file browsing.',
-          label: 'Default workspace',
-          noDefault: 'No default workspace',
-        },
-        searchBackend: {
-          title: 'Web Search Backend',
-          description: 'Choose how agent sessions access web search and content retrieval tools.',
-          activeBackend: 'Active backend',
-          autoStart: 'Auto-start MCP servers',
-          autoStartHint: 'Selected MCP servers will start automatically for every task session.',
-          native: 'Claude Native',
-          kindlyWebSearch: 'Kindly Web Search',
-          ccWebMcp: 'CC-Web-MCP',
-          saving: 'Saving…',
-          saved: 'Saved',
-        },
-        version: {
-          title: 'Deployment Versions',
-          description: 'Git commits baked into the running backend and frontend builds.',
-          backendLabel: 'Backend',
-          frontendLabel: 'Frontend',
-          commitLabel: 'Commit',
-          committedAtLabel: 'Committed at',
-          unavailable: 'Unavailable',
-          mismatchWarning: 'Backend and frontend were built from different commits.',
-        },
-        account: {
-          title: 'Account & Security',
-          description: 'Manage your account credentials.',
-        },
-        tabs: {
-          general: 'General',
-          llmProviders: 'LLM Providers',
-          monitoring: 'Monitoring & Audit',
-          users: 'User Management',
-          envAccess: 'Environment Access',
-          collaborators: 'Collaborators',
-        },
-        monitoring: {
-          title: 'Observability Platforms',
-          description: 'Quick access to monitoring, metrics, and LLM observability dashboards.',
-          noServices: 'No monitoring services are configured. Set the appropriate environment variables on the backend to enable links.',
-          notConfigured: 'Not configured',
-          openService: 'Open {{name}}',
-          grafanaDesc: 'Metrics dashboards, alerts, and visualization',
-          prometheusDesc: 'Time-series metrics collection and querying',
-          litefuseDesc: 'LLM observability: traces, generations, and token analytics',
-        },
-        envAccess: {
-          selectEnv: 'Select environment...',
-          maxTasks: 'Max tasks',
-          unlimited: 'unlimited',
-          grant: 'Grant',
-          remove: 'Remove',
-          grantTo: 'Grant to...',
-          noEnvSelected: 'Select an environment to manage access.',
-          noAccess: 'No access grants yet.',
-        },
-        collaborators: {
-          selectProject: 'Select project...',
-          add: 'Add',
-          addUser: 'Add user...',
-          remove: 'Remove',
-          role: {
-            member: 'Member',
-            viewer: 'Viewer',
-          },
-          noProjectSelected: 'Select a project to manage collaborators.',
-          noCollaborators: 'No collaborators yet.',
-          loadError: 'Failed to load collaborators',
-        },
-        engine: {
-          claudeCode: 'Claude Code',
-          claudeAgent: 'Agent SDK',
-          codexAppServer: 'Codex App Server',
-        },
-        codex: {
-          model: 'Model',
-          command: 'Command',
-          approval: 'Approval',
-          baseUrl: 'Base URL',
-          apiKey: 'API Key',
-          config: 'Codex config.toml',
-          auth: 'Codex auth.json',
-          install: 'Install {{name}}',
-          installing: 'Installing {{name}}…',
-          update: 'Update {{name}}',
-          updating: 'Updating…',
-          installed: '{{name}} installed',
-          updateTitle: 'Update {{name}}',
-          updateWarning: 'This will replace existing files. Continue?',
-          forceUpdate: 'Force Update',
-          placeholders: {
-            baseUrl: 'https://api.openai.com/',
-            apiKey: 'sk-...',
-            model: 'gpt-5-codex',
-            command: 'codex app-server --listen stdio://',
-            approval: 'never',
-            config: 'model = "gpt-5-codex"',
-            auth: '{\n  "OPENAI_API_KEY": "sk-..."\n}',
-          },
-        },
-        appearance: {
-          title: 'Appearance',
-          description: 'Global font family preference',
-          fontFamilyLabel: 'Font family',
-          sansSerif: 'Sans-serif',
-          serif: 'Serif',
-          previewHint: 'Changes apply immediately across the entire interface.',
-        },
-        users: {
-          title: 'User Management',
-          count: '{{count}} users',
-          approve: 'Approve',
-          disable: 'Disable',
-          reEnable: 'Re-enable',
-          resetPassword: 'Reset Password',
-          newPassword: 'New Password:',
-          enterNewPassword: 'Enter new password',
-          set: 'Set',
-        },
-        llmProviders: {
-          title: 'LLM Providers',
-          description: 'Manage global LLM API configurations for quick profile fill.',
-          addProvider: 'Add Provider',
-          editProvider: 'Edit Provider',
-          deleteProvider: 'Delete Provider',
-          noProviders: 'No providers configured yet. Add one to enable quick fill in agent profiles.',
-          nameLabel: 'Name',
-          formatLabel: 'Format',
-          baseUrlLabel: 'Base URL',
-          apiKeyLabel: 'API Key',
-          opusModelLabel: 'Opus Model',
-          sonnetModelLabel: 'Sonnet Model',
-          haikuModelLabel: 'Haiku Model',
-          defaultModelLabel: 'Default Model',
-          formatAnthropic: 'Anthropic',
-          formatOpenAIChat: 'OpenAI Chat',
-          formatOpenAIResponses: 'OpenAI Responses',
-          confirmDelete: 'Delete provider "{{name}}"?',
-          fillFromProvider: 'Fill from provider',
-          customOption: 'Custom',
-          placeholders: {
-            name: 'e.g. Kimi Coding',
-            anthropicBaseUrl: 'https://api.anthropic.com/',
-            openAiBaseUrl: 'https://api.openai.com/',
-            apiKey: 'sk-...',
-            opusModel: 'claude-opus-4-7',
-            sonnetModel: 'claude-sonnet-4-6',
-            haikuModel: 'claude-haiku-4-5',
-            defaultModel: 'gpt-4o',
-          },
-          updateFailed: 'Update failed',
-        },
-      },
-      placeholder: {
-        eyebrow: 'Settings',
-        title: 'Settings are coming soon',
-        description:
-          'This area will host environment preferences, runtime configuration, and WebUI behavior controls in a later slice.',
-        badge: 'Placeholder section',
-      },
-      environments: {
-        eyebrow: '容器',
-        title: 'Containers',
-        description:
-          'Add and manage SSH-backed compute environments. Run detection to verify connectivity, then mark one as active for terminals and tasks.',
-        currentSelection: 'Current selection',
-        activeSelectionFallback: 'No active environment selected yet',
-        addEnvironment: 'Add environment',
-        listTitle: 'Environment list',
-        listDescription:
-          'Use the list to inspect the current control-plane state, run detection, or mark an environment active for downstream pages.',
-        loading: 'Loading environments...',
-        empty: 'No environments have been created yet.',
-        alias: 'Alias',
-        host: 'Host',
-        auth: 'Auth',
-        detection: 'Detection',
-        actions: 'Actions',
-        defaultWorkdir: 'No default workdir',
-        activeBadge: 'Active',
-        editingBadge: 'Editing',
-        detectedAt: 'Detected',
-        notDetected: 'Not detected yet',
-        confirmDelete: 'Delete environment "{{alias}}"?',
-        defaultEnvironmentLocked: 'The default localhost environment cannot be deleted.',
-        detectFallbackToast: '{{alias}} detection fell back to personal tmux after SSH was unavailable.',
-        detectFailedToast: 'Environment detection failed.',
-        detectionStatus: {
-          success: 'Success',
-          partial: 'Partial',
-          failed: 'Failed',
-        },
-        authKind: {
-          ssh_key: 'SSH key',
-          password: 'Password',
-          agent: 'Agent',
-        },
-        projectReferenceTitle: 'Project environment binding',
-        projectReferenceDescription:
-          'Attach the selected environment to the current project, mark it as the project default, and override runtime-only fields such as workdir or env manager.',
-        projectReferenceNoSelection: 'Select an environment from the list before editing project bindings.',
-        projectReferencedBadge: 'Project ref',
-        projectUnreferencedBadge: 'Not referenced',
-        projectDefaultBadge: 'Project default',
-        setProjectDefault: 'Default',
-        setProjectDefaultDisabled: 'Is default',
-        useTooltip: 'Use as session default',
-        defaultTooltip: 'Set as new project default',
-        clearProjectDefault: 'Clear project default',
-        removeProjectReference: 'Remove project ref',
-        projectReferenceSaveError: 'Unable to save project environment binding',
-        projectOverrideWorkdir: 'Override workdir',
-        projectOverrideWorkdirPlaceholder: '/workspace/project-override',
-        projectOverrideEnvName: 'Override env name',
-        projectOverrideEnvNamePlaceholder: 'research-env',
-        projectOverrideEnvManager: 'Override env manager',
-        projectOverrideEnvManagerPlaceholder: 'conda',
-        projectOverrideRuntimeNotes: 'Override runtime notes',
-        projectOverrideRuntimeNotesPlaceholder: 'Prefer the project-specific runtime image.',
-        projectReferenceSaving: 'Saving binding…',
-        updateProjectReference: 'Update project binding',
-        attachProjectReference: 'Attach to project',
-      },
-      resources: {
-        eyebrow: 'RESOURCES',
-        title: 'Resource Monitor',
-        description: 'Real-time system resource usage across all environments.',
-        loading: 'Loading resources...',
-        noData: 'No resource data available yet.',
-        systemCard: {
-          title: 'System Resources',
-          gpuTitle: 'GPU',
-          cpuTitle: 'CPU',
-          memoryTitle: 'Memory',
-          noGpu: 'No GPU detected',
-        },
-        processCard: {
-          title: 'AINRF Processes',
-          empty: 'No AINRF processes detected',
-          columns: {
-            pid: 'PID',
-            name: 'Name',
-            cpu: 'CPU%',
-            memory: 'Memory',
-            runtime: 'Runtime',
-          },
-        },
-        taskUsage: {
-          title: 'Task Usage',
-          description: 'Token and duration totals across all tasks.',
-          taskCount: '{{count}} tasks',
-          totalTokens: 'Total tokens',
-          totalDuration: 'Total task duration',
-          medianDuration: 'Median task duration',
-          topTasks: 'Top token tasks',
-          tokens: '{{count}} tokens',
-          noTopTasks: 'No task token usage recorded yet.',
-        },
-        status: {
-          ok: 'Healthy',
-          degraded: 'Degraded',
-          unavailable: 'Unavailable',
-        },
-      },
-      sessions: {
-        sidebarTitle: 'Task runs',
-        sidebarCount: '{{count}} runs',
-        searchPlaceholder: 'Search task runs...',
-        empty: 'No task runs yet',
-        selectPrompt: 'Select a task run to view details',
-        notFound: 'Task run not found',
-        outputCount: '{{count}} output events',
-        taskCount: '{{count}} tasks',
-        attemptsTitle: 'Attempts',
-        noAttempts: 'No attempts recorded',
-        attemptLabel: 'Attempt #{{seq}}',
-        viewTask: 'View task',
-        hasTokens: 'Token data',
-        status: {
-          active: 'Active',
-          completed: 'Completed',
-          archived: 'Archived',
-        },
-        attemptStatus: {
-          running: 'Running',
-          completed: 'Completed',
-          failed: 'Failed',
-          interrupted: 'Interrupted',
-        },
-        perModelBreakdown: 'Per-Model Breakdown',
-        detail: {
-          prompt: 'Prompt',
-          engine: 'Engine',
-          project: 'Project',
-          workspace: 'Workspace',
-          environment: 'Environment',
-          workdir: 'Workdir',
-          command: 'Command',
-          started: 'Started',
-          completed: 'Completed',
-          outputSeq: 'Output events',
-          exitCode: 'Exit code',
-        },
-        totalTokens: '{{count}} tokens',
-        fileBrowser: {
-          emptyDirectory: 'Empty directory',
-          loading: 'Loading files...',
-          files: 'Files',
-          noFileSelected: 'No file selected',
-          noFiles: 'No files',
-          loadingFile: 'Loading file...',
-          selectEnv: 'Select an environment to browse files',
-          refresh: 'Refresh',
-          selectFile: 'Select a file to view its contents',
-          binaryFile: 'Binary file',
-        },
-        timeline: {
-          title: 'Task run timeline',
-          empty: 'No task runs in this time range',
-          loading: 'Loading task timeline...',
-          taskCount: '{{count}} task runs',
-          projectFilter: 'Filter task timeline by project',
-          allProjects: 'All Projects',
-          today: 'Today',
-          past7Days: 'Past 7 days',
-          past30Days: 'Past 30 days',
-          from: 'From',
-          to: 'To',
-          attemptSeq: 'Attempt #{seq}',
-          status: 'Status:',
-          duration: 'Duration:',
-          reason: 'Reason:',
-          attemptLabel: 'Attempt #{{seq}}',
-          viewTask: 'View task',
-          viewSession: 'View session',
-        },
-      },
-    },
-    components: {
-      healthStatusBar: {
-        loading: 'Loading system status...',
-        unable: 'Unable to fetch system status',
-        apiLabel: 'API:',
-        containerLabel: 'Container:',
-        sshLabel: 'SSH:',
-        workspaceLabel: 'Workspace:',
-      },
-      environmentSelector: {
-        eyebrow: 'Environment selector',
-        title: 'Shared environment selection',
-        description:
-          'Pick the execution environment once and reuse it across the frontend. Selection priority is: settings project default, remembered local choice, localhost seed, then the first available environment.',
-        selectedNone: 'No environment selected',
-        activeLabel: 'Active environment',
-        noAvailable: 'No environments available',
-        selectionStatus: 'Selection status',
-        loaded: 'Loaded:',
-        available: 'Available:',
-        loadError: 'Unable to load environments:',
-        displayName: 'Display name:',
-        host: 'Host:',
-        defaultWorkdir: 'Default workdir:',
-        waitingValidSelection: 'Waiting for a valid selection.',
-        createPrompt:
-          'Create an environment in Containers to enable terminal and workspace selection.',
-      },
-      terminalBench: {
-        eyebrow: '个人终端会话',
-        title: 'Personal terminal session',
-        description:
-          'Attach, detach, or reset the persistent personal terminal session for the selected environment, then open the interactive browser console.',
-        statusPrefix: 'Status:',
-        sessionSource: 'Session source',
-        loading: 'Loading:',
-        websocketUrl: 'WebSocket URL:',
-        environment: 'Environment:',
-        sessionName: 'Session name:',
-        attachment: 'Attachment:',
-        loadError: 'Load error:',
-        selectEnvironmentBeforeAttach: 'Select an environment before attaching the terminal session.',
-        noSessionYet: 'No personal terminal session has been created yet.',
-        detachedNotice: 'The personal session is still running, but this page is currently detached.',
-        attach: 'Attach',
-        detach: 'Detach',
-        resetSession: 'Reset session',
-        attaching: 'Attaching…',
-        detaching: 'Detaching…',
-        resetting: 'Resetting…',
-        consoleTitle: 'Terminal Console',
-      },
-      terminalConsole: {
-        startPrompt: 'Attach to a personal terminal session to open the xterm console.',
-        writeEnabled: 'Write enabled',
-        websocketSession: 'WebSocket session:',
-        attachment: 'Attachment:',
-        observeOnly: 'Observe-only',
-        connection: 'Connection:',
-        exited: 'Session exited with code {{code}}',
-      },
-      environmentEditor: {
-        eyebrow: 'Environment editor',
-        createTitle: 'Create environment',
-        editTitle: 'Edit environment',
-        description:
-          'Define SSH target details, runtime preferences, and the minimal metadata that powers the Containers control plane.',
-        activeLabel: 'Active environment:',
-        noActive: 'No environment has been marked as active yet.',
-        alias: 'Alias',
-        displayName: 'Display name',
-        descriptionField: 'Description',
-        host: 'Host',
-        port: 'Port',
-        user: 'User',
-        authKindLabel: 'Auth kind',
-        identityFile: 'Identity file',
-        proxyJump: 'Proxy jump',
-        proxyCommand: 'Proxy command',
-        tags: 'Tags',
-        defaultWorkdir: 'Default workdir',
-        preferredPython: 'Preferred Python',
-        preferredEnvManager: 'Preferred env manager',
-        preferredRuntimeNotes: 'Preferred runtime notes',
-        sshOptionsJson: 'SSH options JSON',
-        placeholders: {
-          alias: 'gpu-lab',
-          displayName: 'GPU Lab',
-          description: 'Primary training environment',
-          host: 'gpu.example.com',
-          port: '22',
-          user: 'root',
-          identityFile: '/keys/gpu-lab',
-          proxyJump: 'bastion',
-          proxyCommand: 'ssh -W %h:%p bastion',
-          tags: 'gpu, research',
-          defaultWorkdir: '/workspace/project',
-          preferredPython: 'python3.13',
-          preferredEnvManager: 'uv',
-          preferredRuntimeNotes: 'Use CUDA 12 image',
-          sshOptionsJson: '{"StrictHostKeyChecking":"no"}',
-        },
-        authKindOptions: {
-          ssh_key: 'SSH Key',
-          password: 'Password',
-          agent: 'Agent',
-        },
-        saveError: 'Unable to save environment',
-        portRangeError: 'Port must be between 1 and 65535',
-        noEnvironmentSelectedForEditing: 'No environment is selected for editing',
-        sshOptionsObjectError: 'SSH options must be a JSON object',
-        sshOptionsValuesError: 'SSH options values must be strings',
-        create: 'Create environment',
-        save: 'Save changes',
-        saving: 'Saving…',
-        reset: 'Reset',
-        cancelEdit: 'Cancel edit',
-      },
-      errorBoundary: {
-        title: 'Something went wrong',
-        retry: 'Try again',
-      },
-      modal: {
-        close: 'Close',
-      },
-      skills: {
-        ungrouped: 'Ungrouped',
-        mixed: 'Mixed',
-      },
-      token: {
-        tokens: 'Tokens',
-        total: 'Total:',
-        input: 'Input',
-        cache: 'Cache',
-        output: 'Output',
-        think: 'Think',
-      },
-      resources: {
-        cpu: 'CPU Usage',
-        cores: 'cores',
-        noGpu: 'No GPU detected',
-        memory: 'Memory',
-      },
-      environmentDetectionModal: {
-        groups: {
-          basicInfo: 'Basic Info',
-          pythonToolchain: 'Python Toolchain',
-          devTools: 'Dev Tools',
-          aiMl: 'AI / ML',
-          gpu: 'GPU Info',
-          envVars: 'Environment Variables',
-        },
-        labels: {
-          ssh: 'SSH',
-          tmux: 'Tmux',
-          hostname: 'Hostname',
-          os: 'OS',
-          arch: 'Architecture',
-          workdir: 'Workdir',
-          python: 'Python',
-          conda: 'Conda',
-          uv: 'uv',
-          pixi: 'pixi',
-          codex: 'Codex',
-          claudeCli: 'Claude CLI',
-          torch: 'PyTorch',
-          cuda: 'CUDA',
-          gpuCount: 'GPU Count',
-          gpuModels: 'GPU Models',
-          anthropic: 'Anthropic',
-        },
-        warningsHeader: 'Warnings',
-        status: {
-          available: 'Available',
-          unavailable: 'Unavailable',
-          present: 'Present',
-          missing: 'Missing',
-          unknown: 'Unknown',
-        },
-      },
-    },
-  },
-  zh: {
-    chat: {
-      thinking: '思考过程',
-      disclaimer: 'AI 可以犯错。请核实重要信息。',
-      inputPlaceholder: '给 AI 发送消息…',
-      send: '发送消息',
-      attach: '附加文件',
-      webSearch: '联网搜索',
-      reason: '深度思考',
-      copy: '复制',
-      retry: '重试',
-      collapseTools: '折叠工具',
-      scrollToBottom: '滚动到底部',
-    },
-    common: {
-      appName: 'AINRF 控制台',
-      language: '语言',
-      english: '英文',
-      chinese: '中文',
-      loading: '加载中...',
-      saving: '保存中…',
-      reset: '重置',
-      cancel: '取消',
-      save: '保存',
-      saveChanges: '保存更改',
-      createEnvironment: '创建环境',
-      use: '使用',
-      edit: '编辑',
-      detect: '探测',
-      delete: '删除',
-      default: '默认',
-      start: '启动',
-      stop: '停止',
-      active: '活动',
-      editing: '编辑中',
-      never: '从未',
-      notConfigured: '未配置',
-      notSelected: '未选择',
-      unavailable: '不可用',
-      available: '可用',
-      writable: '可写',
-      notReady: '未就绪',
-      configured: '已配置',
-      detailLabel: '详情：',
-      noTags: '无标签',
-      noDefaultWorkdir: '无默认工作目录',
-      yes: '是',
-      no: '否',
-      ok: 'OK',
-      degraded: '降级',
-      idle: '空闲',
-      starting: '启动中',
-      running: '运行中',
-      stopping: '停止中',
-      failed: '失败',
-      connecting: '连接中',
-      connected: '已连接',
-      disconnected: '已断开',
-      error: '错误',
-      builtBy: '由 Kozmosa 用 ❤️ 构建',
-      collapse: '折叠',
-      confirmLogout: '确认登出',
-      confirmLogoutMessage: '确定要登出吗？',
-      dragToReorder: '拖拽排序',
-      expand: '展开',
-      logOut: '登出',
-      taskStatusLoading: '任务 | 加载中…',
-      taskStatusSummary: '任务 | 共 {{total}} 个，运行中：{{running}}，等待中：{{pending}}，已完成：{{finished}}',
-      taskStatusUnavailable: '任务 | 状态不可用',
-      loadingEditor: '加载编辑器...',
-      noMessages: '暂无消息',
-    },
-    nav: {
-      literature: '文献追踪',
-    },
-    literature: {
-      mySubscriptions: '我的订阅',
-      newSubscription: '新建订阅',
-      label: '标签',
-      keywords: '关键词（逗号分隔）',
-      keywordsPlaceholder: '机器学习, NLP',
-      categories: '分类',
-      frequency: '频率',
-      daily: '每日',
-      weekly: '每周',
-      twice_daily: '每日两次',
-      create: '创建',
-      delete: '删除',
-      unreadOnly: '仅未读',
-      refresh: '刷新',
-      markRead: '标记已读',
-      viewarXiv: 'arXiv',
-      convertToTask: '转换为任务',
-      convertConfirm: '将此论文转换为 Agent 任务。请在下方选择工作区、运行环境和执行引擎。',
-      convertSuccess: '论文已成功转换为任务',
-      convertError: '转换论文为任务失败',
-      noPapers: '暂无论文。请先创建订阅。',
-      noSubscriptions: '暂无订阅。创建一个订阅以追踪论文。',
-      aiPracticeNote: 'AI 实践提醒',
-      aiSummary: 'AI 摘要',
-    },
-    auth: {
-      login: '登录',
-      logout: '退出',
-      username: '用户名',
-      usernameHint: '仅限字母、数字、点、下划线、连字符',
-      password: '密码',
-      displayName: '显示名称',
-      confirmPassword: '确认密码',
-      passwordMismatch: '密码不匹配',
-      registerFailed: '注册失败',
-      registrationSubmitted: '注册已提交',
-      backToLogin: '返回登录',
-      registerLink: '没有账号？注册',
-      loginLink: '已有账号？登录',
-      register: '注册',
-      pendingApproval: '账号等待管理员审批中。',
-      changePassword: '修改密码',
-      mustChangePassword: '首次登录需要修改密码。',
-      currentPassword: '当前密码',
-      newPassword: '新密码',
-      changePasswordFailed: '修改密码失败',
-      passwordTooShort: '密码至少需要4个字符',
-      loginFailed: '登录失败',
-    },
-    navigation: {
-      projects: {
-        label: '项目',
-        description: '项目画布与任务图可视化',
-      },
-      terminal: {
-        label: '终端',
-        description: '健康检查和终端工作台入口',
-      },
-      tasks: {
-        label: '任务',
-        description: 'Agent 窗口与只读观察终端',
-      },
-      workspaces: {
-        label: '工作区',
-        description: '项目工作区和任务上下文',
-      },
-      workspaceBrowser: {
-        label: '工作区浏览',
-        description: '浏览工作区文件',
-      },
-      environments: {
-        label: '容器',
-        description: '运行环境与容器状态',
-      },
-      resources: {
-        label: '资源监控',
-        description: '系统资源占用与使用情况',
-      },
-      sessions: {
-        label: '任务运行',
-        description: '来自真实 Agent 任务的运行历史',
-      },
-      timeline: {
-        label: '任务时间线',
-        description: '来自真实 Agent 任务的运行时间线',
-      },
-      settings: {
-        label: '设置',
-        description: '运行时和 WebUI 偏好设置',
-      },
-    },
-    layout: {
-      brandLine: '与后端健康检查、终端工作台、环境控制平面和托管工作区浏览器对齐的轻量前端壳。',
-      headerEyebrow: 'WebUI 控制面',
-      headerDescription: '后端重整期间的健康、终端工作台、环境控制平面和工作区浏览器外壳。',
-      footerDescription: '后端重整期间的健康、终端工作台、环境控制平面和工作区浏览器外壳。',
-      collapseSidebar: '收起侧边栏',
-      expandSidebar: '展开侧边栏',
-      resizeSidebar: '调整侧边栏宽度',
-    },
-    pages: {
-      projects: {
-        sidebarEyebrow: '项目',
-        sidebarTitle: '项目画布',
-        sidebarCount: '{{count}} 个项目',
-        newProject: '新建项目',
-        searchPlaceholder: '搜索项目',
-        newTask: '新建任务',
-        resetLayout: '重置布局',
-        emptyCanvas: "点击'新建任务'开始",
-        noProjects: '暂无项目。创建一个项目开始。',
-        moveToProject: '移动到项目',
-        noOtherProjects: '没有其他可移动到的项目。',
-        createTitle: '创建项目',
-        createNameLabel: '项目名称',
-        createDescriptionLabel: '描述（可选）',
-        createSubmit: '创建项目',
-      },
-      dashboard: {
-        eyebrow: '终端',
-        title: 'Scholar Agent 运行时外壳',
-        description:
-          '前端现在暴露了整理后的运行时外壳：健康检查、单会话终端工作台控制面、环境控制平面和托管工作区浏览面板。此页用于确认 API、SSH、工作区就绪状态、终端会话状态和环境就绪状态。',
-        surfaceTitle: '当前后端面',
-        endpointsLabel: '可用端点：',
-        stateRootLabel: '状态根：',
-        detailLabel: '详情：',
-      },
-      terminal: {
-        eyebrow: 'TERMINAL',
-        title: '终端',
-        description: '',
-        backToTasks: '返回任务页',
-        task: {
-          observeOnly: '仅观察',
-          takenOver: '已接管',
-          archived: '已归档',
-          pending: '等待附着',
-          windowLabel: '窗口：',
-          ownerLabel: '所有者：',
-          environmentLabel: '环境：',
-          noOwner: '当前无人占用',
-          lastOutputLabel: '最近输出：',
-          otherOwner: '该任务当前由其他浏览器用户接管，此处无法释放。',
-          placeholder: '任务终端附着尚未激活。',
-        },
-      },
-      tasks: {
-        sidebarEyebrow: '任务运行',
-        sidebarTitle: 'Agent 任务',
-        sidebarCount: '共 {{count}} 个 · 实时输出回放',
-        newTask: '新建任务',
-        resizeTaskList: '调整任务列表宽度',
-        searchLabel: '搜索任务',
-        searchPlaceholder: '搜索任务',
-        empty: '还没有任务。新建一个运行后即可开始。',
-        noSearchResults: '没有匹配“{{query}}”的任务。',
-        updatedAt: '更新于 {{time}}',
-        createTitle: '创建任务',
-        createDescription: '在工作区与环境绑定上启动一个 Claude Code 任务。',
-        closeCreate: '关闭创建任务',
-        workspaceLabel: '工作区',
-        environmentLabel: '环境',
-        projectLabel: '项目',
-        profileLabel: '任务配置',
-        researchAgentLabel: '研究代理',
-        taskConfigurationLabel: '任务配置模式',
-        skillsLabel: '技能',
-        skillDependenciesHint: '有依赖的技能将在运行时自动包含所需技能。',
-        titleLabel: '标题',
-        optionalPlaceholder: '可选',
-        taskInputLabel: '任务输入',
-        taskInputPlaceholder: '描述要交给 Agent 的任务。',
-        structuredResearchGoal: '研究目标',
-        structuredResearchContext: '上下文',
-        structuredResearchConstraints: '约束',
-        structuredResearchDeliverables: '预期交付物',
-        structuredResearchValidationPlan: '验证计划',
-        generatedPromptPreview: '生成提示词预览',
-        paperPathLabel: '论文路径',
-        paperPathPlaceholder: 'papers/target.pdf',
-        scopeLabel: '范围',
-        targetTableLabel: '目标表格（可选）',
-        targetTablePlaceholder: 'Table3',
-        budgetHoursLabel: '预算（小时）',
-        researchTopicLabel: '研究主题',
-        researchTopicPlaceholder: '用于药物发现的图神经网络',
-        seedPaperPathLabel: '种子论文路径（可选）',
-        seedPaperPathPlaceholder: 'papers/seed.pdf',
-        depthLabel: '深度',
-        ideaSourceLabel: '想法来源',
-        ideaSourcePlaceholder: '工作区中的想法文档路径，或在此描述你的想法',
-        validationScopeLabel: '验证范围',
-        scopeCoreOnly: '核心方法',
-        scopeFullSuite: '完整套件',
-        validationScopeQuick: '快速',
-        validationScopeFull: '完整',
-        defaultWorkdir: '默认工作目录：',
-        resetDraft: '重置草稿',
-        createAction: '创建任务',
-        creatingAction: '创建中…',
-        create: {
-          taskPreset: '任务预设',
-          presets: {
-            rawPrompt: '原始提示词',
-            structuredResearch: '结构化研究',
-            reproduceBaseline: '复现实验基线',
-            overview: '概要总结',
-          },
-          researcherType: '研究员类型',
-          researcherVanilla: 'Vanilla',
-          researcherAris: 'ARIS Researcher',
-          executionEngine: '执行引擎',
-          engineClaudeCode: 'Claude Code',
-          engineAgentSdk: 'Agent SDK',
-          engineCodexAppServer: 'Codex App Server',
-          skillsPlaceholder: 'analysis, code-review',
-          titlePlaceholder: '可选任务标题',
-          promptLabel: 'Prompt',
-          promptPlaceholder: '输入研究任务提示词…',
-          missingBinding: '请先选择工作区和环境再创建任务。',
-          noSkillsAvailable: '当前没有可用于此任务的技能。',
-          skillGroupSelected: '已选择 {{selected}}/{{total}}',
-          showSkillGroup: '显示 {{group}} 中的技能',
-          hideSkillGroup: '隐藏 {{group}} 中的技能',
-          selectSkill: '选择 {{skill}}',
-          deselectSkill: '取消选择 {{skill}}',
-          seedFiles: {
-            dropzone: '拖拽种子论文到此处，或点击选择文件（支持 PDF、Markdown）',
-            uploading: '正在上传 {{name}}…',
-            uploadError: '上传 {{name}} 失败：{{error}}',
-            removeFile: '移除 {{name}}',
-            unsupportedType: '仅支持 PDF 和 Markdown 文件。',
-            maxFilesExceeded: '最多上传 {{max}} 篇种子论文。',
-            fileCount: '已上传 {{count}} 篇种子论文',
-            label: '种子论文',
-          },
-        },
-        noTaskSelected: '未选择任务',
-        noTaskSelectedDescription: '从侧边栏选择一个任务，或新建一次运行来查看输出。',
-        workspaceEyebrow: '任务工作区',
-        outputTimeline: '输出时间线',
-        latestSeq: '最新序号 {{seq}}',
-        noOutput: '尚未记录任何输出。',
-        promptLayers: 'Prompt 分层',
-        prompt: 'Prompt',
-        chars: '字符',
-        promptUnavailable: 'Prompt 清单尚不可用。',
-        summary: '摘要',
-        taskId: '任务 ID',
-        created: '创建时间',
-        updated: '更新时间',
-        started: '开始时间',
-        completed: '完成时间',
-        binding: '绑定',
-        profile: '配置',
-        workdir: '工作目录',
-        snapshot: '快照',
-        bindingUnavailable: '绑定快照尚不可用。',
-        runtime: '运行时',
-        runner: 'Runner',
-        directory: '目录',
-        command: '命令',
-        runtimeUnavailable: '运行时载荷尚不可用。',
-        result: '结果',
-        exitCode: '退出码',
-        failure: '失败',
-        unavailable: '不可用',
-        pendingValue: '等待中',
-        status: {
-          queued: '排队中',
-          starting: '启动中',
-          running: '运行中',
-          succeeded: '已成功',
-          failed: '已失败',
-          cancelled: '已取消',
-          paused: '已暂停',
-        },
-        expandOutputTimeline: '展开输出时间线',
-        collapseOutputTimeline: '收起输出时间线',
-        expandSummary: '展开摘要面板',
-        collapseSummary: '收起摘要面板',
-        thinking: '思考中…',
-        toolCall: '工具：{{name}}',
-        toolResult: '结果',
-        inputPlaceholder: '输入消息…',
-        metadata: {
-          workdir: '工作目录',
-          command: '命令',
-          taskInput: '任务输入',
-        },
-        layout: {
-          showOnlyDetails: '仅显示详情',
-          showOnlyConversation: '仅显示对话',
-          collapseSidebar: '收起详情',
-          expandSidebar: '展开详情',
-          resizeSidebar: '调整详情面板宽度',
-          errorIndicator: '任务有错误 — 点击查看',
-        },
-        output: {
-          replayFailed: '无法回放任务输出',
-          parseFailed: '无法解析任务输出',
-          loadFailed: '无法加载任务输出',
-        },
-        messageGroup: {
-          summary: '{{count}} 条{{type}}消息',
-        },
-        messageType: {
-          systemEvent: '系统',
-          user: '用户',
-          assistant: '助手',
-          thinking: '思考',
-          toolCall: '工具调用',
-          toolResult: '工具结果',
-          stdout: '标准输出',
-          stderr: '标准错误',
-          lifecycle: '生命周期',
-        },
-        actions: {
-          cancel: '取消',
-          archive: '归档',
-          retry: '重试',
-          showArchived: '显示已归档',
-          pause: '暂停',
-          resume: '继续',
-          pauseFailed: '暂停失败：{{error}}',
-          resumeFailed: '继续失败：{{error}}',
-          sendPromptFailed: '发送提示词失败：{{error}}',
-          noTaskSelected: '未选择任务',
-          unexpectedError: '发生意外错误',
-        },
-        retrySuccess: '任务已重试',
-        retryFailed: '重试失败',
-        sort: {
-          updated: '更新时间（从新到旧）',
-          created: '创建时间（从新到旧）',
-          name: '名称（A-Z）',
-        },
-      },
-      workspaces: {
-        eyebrow: 'WORKSPACES',
-        title: '工作区',
-        description:
-          '管理可复用的工作区上下文，用于任务提示词、默认工作目录和研究运行边界。',
-        managerTitle: '工作区管理器',
-        managerDescription: '创建和维护任务启动前可绑定的工作区记录。',
-        newWorkspace: '新建工作区',
-        createWorkspace: '创建工作区',
-        saveWorkspace: '保存工作区',
-        deleteWorkspace: '删除工作区',
-        confirmDelete: '确认删除',
-        labelField: '工作区名称',
-        descriptionField: '说明',
-        defaultWorkdirField: '默认工作目录',
-        promptField: '工作区提示词',
-        noDefaultWorkdir: '未配置默认工作目录',
-      },
-      workspaceBrowser: {
-        eyebrow: 'WORKSPACE BROWSER',
-        title: '工作区浏览器',
-        description: '打开设置页所选环境的托管 code-server 工作区浏览器。',
-        readinessTitle: '需要完成运行时设置',
-        readinessMissingDependency: '缺少运行时依赖。',
-      },
-      settings: {
-        eyebrow: 'SETTINGS',
-        title: '设置',
-        description:
-          '自定义 Scholar Agent 的使用体验：默认首页、终端字号、项目默认环境以及任务草稿模板。',
-        recoveryNotice:
-          '本地设置文档缺字段、已损坏或版本不兼容，应用已回退到默认配置以继续运行。',
-        routes: {
-          terminal: '终端',
-          tasks: '任务',
-          workspaces: '工作区',
-          environments: '环境',
-        },
-        general: {
-          title: '通用偏好',
-          description: '作用于整个界面的偏好设置。',
-          defaultRouteLabel: '默认首页',
-          terminalFontSizeLabel: '终端字号',
-          terminalFontSizeHelp:
-            '字号会被限制在稳定区间 {{min}}-{{max}} px，当前保存值为 {{current}} px。',
-          editorFontSizeLabel: '编辑器字号',
-          editorFontFamilyLabel: '编辑器字体',
-          editorFontSizeHelp:
-            '编辑器字号会被限制在 {{min}}-{{max}} px，当前保存值为 {{current}} px。',
-        },
-        project: {
-          title: '项目默认值',
-          description: '当前项目的默认配置，仅保存在当前浏览器中。',
-          defaultEnvironmentLabel: '默认环境',
-          defaultEnvironmentEmpty: '不设置项目默认环境',
-          defaultEnvironmentHelp:
-            '这个环境优先级高于本地记忆的最近一次显式选择。',
-          noEnvironments:
-            '请先在“容器”页面创建至少一个环境，再配置项目默认值。',
-          environmentCardDescription: '这些字段只用于预填当前环境下的新建任务草稿。',
-          titleTemplateLabel: '标题模板',
-          titleTemplatePlaceholder: '可选的任务标题预设',
-          taskInputTemplateLabel: '任务输入模板',
-          taskInputTemplatePlaceholder: '填写该环境下默认预载入的任务输入。',
-          researchAgentDefaultLabel: '默认研究代理',
-          taskConfigurationDefaultLabel: '默认任务配置',
-        },
-        taskConfiguration: {
-          title: '任务配置',
-          description:
-            '配置创建任务时的三层结构：Claude Code 执行引擎、Research Agent 配置和任务输入模式。',
-          executionEngineLabel: '执行引擎',
-          defaultResearchAgentLabel: '默认研究代理',
-          defaultTaskConfigurationLabel: '默认任务配置',
-          profileLabel: 'Research Agent 标签',
-          systemPromptLabel: '系统提示词',
-          skillsLabel: '启用的技能',
-          skillsDescription: '选择该研究代理应使用的技能。',
-          skillsPromptLabel: 'Skills / 配置说明',
-          settingsJsonLabel: 'Claude settings JSON',
-          apiBaseUrlLabel: 'API Base URL',
-          apiKeyLabel: 'API Key',
-          defaultOpusModelLabel: 'Opus 模型',
-          defaultSonnetModelLabel: 'Sonnet 模型',
-          defaultHaikuModelLabel: 'Haiku 模型',
-          envOverridesLabel: '环境变量覆盖 (JSON)',
-          placeholders: {
-            apiBaseUrl: 'https://api.anthropic.com/',
-            apiKey: 'sk-...',
-            opusModel: 'claude-opus-4-7',
-            sonnetModel: 'claude-sonnet-4-6',
-            haikuModel: 'claude-haiku-4-5',
-            envOverrides: '{\n  "CUSTOM_VAR": "value"\n}',
-          },
-        },
-        skillRepository: {
-          title: 'Skill 仓库',
-          description: '管理 AINRF 技能：查看详情、预览设置并导入新技能。',
-          selectSkill: '选择一个技能以查看详情',
-          version: '版本',
-          author: '作者',
-          dependencies: '依赖',
-          mcpServers: 'MCP 服务器',
-          hooks: '钩子',
-          allowedAgents: '允许的 Agent',
-          skillMdTitle: 'SKILL.md',
-          previewSettings: '预览设置',
-          settingsPreviewTitle: '设置预览',
-          importSkill: '导入技能',
-          importTitle: '导入技能',
-          sourceLabel: '来源',
-          gitSource: 'Git URL',
-          localSource: '本地路径',
-          urlLabel: 'Git URL',
-          pathLabel: '本地路径',
-          skillIdOverrideLabel: 'Skill ID 覆盖（可选）',
-          importAction: '导入',
-          importing: '导入中…',
-          cancel: '取消',
-          noSkills: '仓库中未找到任何技能。',
-          placeholders: {
-            url: 'https://github.com/example/skill-repo.git',
-            path: '/path/to/skill',
-          },
-        },
-        defaultWorkspace: {
-          title: '默认工作区',
-          description: '选择默认工作区用于任务创建和文件浏览。',
-          label: '默认工作区',
-          noDefault: '无默认工作区',
-        },
-        searchBackend: {
-          title: '网络搜索后端',
-          description: '选择 Agent 会话使用的网页搜索和内容抓取工具。',
-          activeBackend: '活跃后端',
-          autoStart: '自动启动 MCP 服务器',
-          autoStartHint: '选中的 MCP 服务器会在每个任务会话中自动启动。',
-          native: 'Claude 原生',
-          kindlyWebSearch: 'Kindly Web Search',
-          ccWebMcp: 'CC-Web-MCP',
-          saving: '保存中…',
-          saved: '已保存',
-        },
-        version: {
-          title: '部署版本',
-          description: '当前运行的后端与前端构建各自内置的 Git 提交。',
-          backendLabel: '后端',
-          frontendLabel: '前端',
-          commitLabel: '提交',
-          committedAtLabel: '提交时间',
-          unavailable: '不可用',
-          mismatchWarning: '后端与前端基于不同的提交构建，版本不一致。',
-        },
-        account: {
-          title: '账号与安全',
-          description: '管理你的账号凭证。',
-        },
-        tabs: {
-          general: '通用',
-          llmProviders: 'LLM 提供商',
-          monitoring: '监控审计',
-          users: '用户管理',
-          envAccess: '环境访问',
-          collaborators: '协作者',
-        },
-        monitoring: {
-          title: '可观测性平台',
-          description: '快速访问监控、指标和 LLM 可观测性仪表盘。',
-          noServices: '未配置任何监控服务。请在后端设置相应的环境变量以启用链接。',
-          notConfigured: '未配置',
-          openService: '打开 {{name}}',
-          grafanaDesc: '指标仪表盘、告警与数据可视化',
-          prometheusDesc: '时序指标采集与查询',
-          litefuseDesc: 'LLM 可观测性：链路追踪、生成记录与 Token 分析',
-        },
-        envAccess: {
-          selectEnv: '选择环境...',
-          maxTasks: '最大任务数',
-          unlimited: '无限制',
-          grant: '授权',
-          remove: '移除',
-          grantTo: '授权给...',
-          noEnvSelected: '选择一个环境以管理访问权限。',
-          noAccess: '暂无授权记录。',
-        },
-        collaborators: {
-          selectProject: '选择项目...',
-          add: '添加',
-          addUser: '添加用户...',
-          remove: '移除',
-          role: {
-            member: '成员',
-            viewer: '观察者',
-          },
-          noProjectSelected: '选择一个项目以管理协作者。',
-          noCollaborators: '暂无协作者。',
-          loadError: '加载协作者失败',
-        },
-        engine: {
-          claudeCode: 'Claude Code',
-          claudeAgent: 'Agent SDK',
-          codexAppServer: 'Codex App Server',
-        },
-        codex: {
-          model: '模型',
-          command: '命令',
-          approval: '审批',
-          baseUrl: 'Base URL',
-          apiKey: 'API Key',
-          config: 'Codex 配置 (TOML)',
-          auth: 'Codex 认证 (JSON)',
-          install: '安装 {{name}}',
-          installing: '正在安装 {{name}}…',
-          update: '更新 {{name}}',
-          updating: '更新中…',
-          installed: '{{name}} 已安装',
-          updateTitle: '更新 {{name}}',
-          updateWarning: '此操作将覆盖现有文件，是否继续？',
-          forceUpdate: '强制更新',
-          placeholders: {
-            baseUrl: 'https://api.openai.com/',
-            apiKey: 'sk-...',
-            model: 'gpt-5-codex',
-            command: 'codex app-server --listen stdio://',
-            approval: 'never',
-            config: 'model = "gpt-5-codex"',
-            auth: '{\n  "OPENAI_API_KEY": "sk-..."\n}',
-          },
-        },
-        appearance: {
-          title: '外观',
-          description: '全局字体偏好设置',
-          fontFamilyLabel: '字体',
-          sansSerif: '非衬线体',
-          serif: '衬线体',
-          previewHint: '更改将立即应用到整个界面。',
-        },
-        users: {
-          title: '用户管理',
-          count: '{{count}} 个用户',
-          approve: '审批通过',
-          disable: '禁用',
-          reEnable: '重新启用',
-          resetPassword: '重置密码',
-          newPassword: '新密码：',
-          enterNewPassword: '输入新密码',
-          set: '设置',
-        },
-        llmProviders: {
-          title: 'LLM 提供商',
-          description: '管理全局 LLM API 配置，以便在 Agent 配置中快速填充。',
-          addProvider: '添加提供商',
-          editProvider: '编辑提供商',
-          deleteProvider: '删除提供商',
-          noProviders: '尚未配置任何提供商。添加一个以在 Agent 配置中启用快速填充。',
-          nameLabel: '名称',
-          formatLabel: '格式',
-          baseUrlLabel: 'Base URL',
-          apiKeyLabel: 'API Key',
-          opusModelLabel: 'Opus 模型',
-          sonnetModelLabel: 'Sonnet 模型',
-          haikuModelLabel: 'Haiku 模型',
-          defaultModelLabel: '默认模型',
-          formatAnthropic: 'Anthropic',
-          formatOpenAIChat: 'OpenAI Chat',
-          formatOpenAIResponses: 'OpenAI Responses',
-          confirmDelete: '删除提供商「{{name}}」？',
-          fillFromProvider: '从提供商填充',
-          customOption: '自定义',
-          placeholders: {
-            name: '例如 Kimi Coding',
-            anthropicBaseUrl: 'https://api.anthropic.com/',
-            openAiBaseUrl: 'https://api.openai.com/',
-            apiKey: 'sk-...',
-            opusModel: 'claude-opus-4-7',
-            sonnetModel: 'claude-sonnet-4-6',
-            haikuModel: 'claude-haiku-4-5',
-            defaultModel: 'gpt-4o',
-          },
-          updateFailed: '更新失败',
-        },
-      },
-      placeholder: {
-        eyebrow: '设置',
-        title: '设置即将推出',
-        description: '此区域将在后续更新中支持环境偏好、运行时配置和行为控制。',
-        badge: '占位区块',
-      },
-      environments: {
-        eyebrow: 'CONTAINERS',
-        title: '容器',
-        description:
-          '添加并管理基于 SSH 的计算环境。运行连通性探测，然后将其中一个标记为活动环境，供终端和任务使用。',
-        currentSelection: '当前选择',
-        activeSelectionFallback: '尚未选择活动环境',
-        addEnvironment: '添加环境',
-        listTitle: '环境列表',
-        listDescription: '选择、编辑、探测或删除环境。活动环境会被高亮显示。',
-        loading: '正在加载环境...',
-        empty: '尚未创建任何环境。',
-        alias: '别名',
-        host: '主机',
-        auth: '认证',
-        detection: '探测',
-        actions: '操作',
-        defaultWorkdir: '未配置默认工作目录',
-        activeBadge: '活动',
-        editingBadge: '编辑中',
-        detectedAt: '探测于',
-        notDetected: '尚未探测',
-        confirmDelete: '删除环境「{{alias}}」？',
-        defaultEnvironmentLocked: '默认的 localhost 环境不能删除。',
-        detectFallbackToast: '{{alias}} 探测在 SSH 不可用后已回退到个人 tmux 会话。',
-        detectFailedToast: '环境探测失败。',
-        detectionStatus: {
-          success: '成功',
-          partial: '部分成功',
-          failed: '失败',
-        },
-        authKind: {
-          ssh_key: 'SSH 密钥',
-          password: '密码',
-          agent: 'SSH Agent',
-        },
-        projectReferenceTitle: '项目环境绑定',
-        projectReferenceDescription:
-          '把当前所选环境绑定到当前项目，设置项目默认环境，并覆盖仅与运行时相关的字段，例如工作目录或环境管理器。',
-        projectReferenceNoSelection: '请先从列表中选择一个环境，再编辑项目绑定。',
-        projectReferencedBadge: '项目引用',
-        projectUnreferencedBadge: '未引用',
-        projectDefaultBadge: '项目默认',
-        setProjectDefault: '默认',
-        setProjectDefaultDisabled: '已是默认',
-        useTooltip: '用作会话默认',
-        defaultTooltip: '设为新项目默认',
-        clearProjectDefault: '取消项目默认',
-        removeProjectReference: '移除项目引用',
-        projectReferenceSaveError: '无法保存项目环境绑定',
-        projectOverrideWorkdir: '覆盖工作目录',
-        projectOverrideWorkdirPlaceholder: '/workspace/project-override',
-        projectOverrideEnvName: '覆盖环境名',
-        projectOverrideEnvNamePlaceholder: 'research-env',
-        projectOverrideEnvManager: '覆盖环境管理器',
-        projectOverrideEnvManagerPlaceholder: 'conda',
-        projectOverrideRuntimeNotes: '覆盖运行时说明',
-        projectOverrideRuntimeNotesPlaceholder: '优先使用项目专用运行时镜像。',
-        projectReferenceSaving: '正在保存绑定…',
-        updateProjectReference: '更新项目绑定',
-        attachProjectReference: '绑定到项目',
-      },
-      resources: {
-        eyebrow: '资源监控',
-        title: '资源监控',
-        description: '所有环境的实时系统资源占用情况。',
-        loading: '正在加载资源数据...',
-        noData: '暂无资源数据。',
-        systemCard: {
-          title: '系统资源',
-          gpuTitle: 'GPU',
-          cpuTitle: 'CPU',
-          memoryTitle: '内存',
-          noGpu: '未检测到 GPU',
-        },
-        processCard: {
-          title: 'AINRF 进程',
-          empty: '未检测到 AINRF 进程',
-          columns: {
-            pid: 'PID',
-            name: '进程名',
-            cpu: 'CPU%',
-            memory: '内存',
-            runtime: '运行时间',
-          },
-        },
-        taskUsage: {
-          title: '任务用量',
-          description: '所有任务的 token 与耗时统计。',
-          taskCount: '{{count}} 个任务',
-          totalTokens: '总 Token',
-          totalDuration: '任务总耗时',
-          medianDuration: '任务耗时中位数',
-          topTasks: 'Token 消耗最高的任务',
-          tokens: '{{count}} 个 token',
-          noTopTasks: '暂无任务 token 用量记录。',
-        },
-        status: {
-          ok: '正常',
-          degraded: '降级',
-          unavailable: '不可用',
-        },
-      },
-      sessions: {
-        sidebarTitle: '任务运行',
-        sidebarCount: '{{count}} 个运行',
-        searchPlaceholder: '搜索任务运行...',
-        empty: '暂无任务运行',
-        selectPrompt: '选择一个任务运行查看详情',
-        notFound: '未找到任务运行',
-        outputCount: '{{count}} 条输出事件',
-        taskCount: '{{count}} 个任务',
-        attemptsTitle: '尝试记录',
-        noAttempts: '暂无尝试记录',
-        attemptLabel: '尝试 #{{seq}}',
-        viewTask: '查看任务',
-        hasTokens: '有 Token 数据',
-        status: {
-          active: '活跃',
-          completed: '已完成',
-          archived: '已归档',
-        },
-        attemptStatus: {
-          running: '运行中',
-          completed: '已完成',
-          failed: '失败',
-          interrupted: '已中断',
-        },
-        perModelBreakdown: '模型明细',
-        detail: {
-          prompt: '提示词',
-          engine: '引擎',
-          project: '项目',
-          workspace: '工作区',
-          environment: '环境',
-          workdir: '工作目录',
-          command: '命令',
-          started: '开始',
-          completed: '完成',
-          outputSeq: '输出事件',
-          exitCode: '退出码',
-        },
-        totalTokens: '{{count}} 个 token',
-        fileBrowser: {
-          loading: '加载文件...',
-          files: '文件',
-          noFileSelected: '未选择文件',
-          noFiles: '无文件',
-          loadingFile: '加载文件...',
-          selectEnv: '选择一个环境以浏览文件',
-          refresh: '刷新',
-          selectFile: '选择一个文件以查看内容',
-          binaryFile: '二进制文件',
-          emptyDirectory: '空目录',
-        },
-        timeline: {
-          title: '任务运行时间线',
-          empty: '此时间段内无任务运行',
-          loading: '正在加载任务时间线...',
-          taskCount: '{{count}} 个任务运行',
-          projectFilter: '按项目筛选任务时间线',
-          allProjects: '全部项目',
-          today: '今天',
-          past7Days: '过去 7 天',
-          past30Days: '过去 30 天',
-          from: '从',
-          to: '至',
-          attemptSeq: '第{seq}次',
-          status: '状态：',
-          duration: '耗时：',
-          reason: '原因：',
-          attemptLabel: '尝试 #{{seq}}',
-          viewTask: '查看任务',
-          viewSession: '查看会话',
-        },
-      },
-    },
-    components: {
-      healthStatusBar: {
-        loading: '正在加载系统状态...',
-        unable: '无法获取系统状态',
-        apiLabel: 'API：',
-        containerLabel: '容器：',
-        sshLabel: 'SSH：',
-        workspaceLabel: '工作区：',
-      },
-      environmentSelector: {
-        eyebrow: '环境选择器',
-        title: '共享环境选择',
-        description:
-          '只需选择一次执行环境，前端各页面都会复用它。优先级固定为设置页中的项目默认环境、本地记忆选择、localhost，最后才是列表首项。',
-        selectedNone: '未选择环境',
-        activeLabel: '当前环境',
-        noAvailable: '没有可用环境',
-        selectionStatus: '选择状态',
-        loaded: '已加载：',
-        available: '可用：',
-        loadError: '无法加载环境：',
-        displayName: '显示名称：',
-        host: '主机：',
-        defaultWorkdir: '默认工作目录：',
-        waitingValidSelection: '等待有效选择。',
-        createPrompt: '请先在“容器”中创建环境，以启用终端和工作区选择。',
-      },
-      terminalBench: {
-        eyebrow: 'PERSONAL TERMINAL SESSION',
-        title: '个人终端会话',
-        description:
-          '连接、脱离或重置当前环境的持久个人终端会话，并在浏览器中打开交互式控制台。',
-        statusPrefix: '状态：',
-        sessionSource: '会话来源',
-        loading: '加载中：',
-        websocketUrl: 'WebSocket URL:',
-        environment: '环境：',
-        sessionName: '会话名：',
-        attachment: '附着 ID：',
-        loadError: '加载错误：',
-        selectEnvironmentBeforeAttach: '附着终端会话前请先选择环境。',
-        noSessionYet: '尚未创建个人终端会话。',
-        detachedNotice: '个人会话仍在运行，但当前页面尚未附着。',
-        attach: '附着',
-        detach: '脱离',
-        resetSession: '重置会话',
-        attaching: '附着中…',
-        detaching: '脱离中…',
-        resetting: '重置中…',
-        consoleTitle: '终端控制台',
-      },
-      terminalConsole: {
-        startPrompt: '附着到个人终端会话后即可打开 xterm 控制台。',
-        writeEnabled: '可写',
-        websocketSession: 'WebSocket 会话：',
-        attachment: '附着：',
-        observeOnly: '仅观察',
-        connection: '连接：',
-        exited: '会话已退出，退出码 {{code}}',
-      },
-      environmentEditor: {
-        eyebrow: '环境编辑器',
-        createTitle: '创建环境',
-        editTitle: '编辑环境',
-        description: '定义 SSH 目标详情、运行时偏好，以及支撑环境控制平面的最小元数据。',
-        activeLabel: '当前活动环境：',
-        noActive: '尚未标记任何活动环境。',
-        alias: '别名',
-        displayName: '显示名称',
-        descriptionField: '描述',
-        host: '主机',
-        port: '端口',
-        user: '用户',
-        authKindLabel: '认证方式',
-        identityFile: '身份文件',
-        proxyJump: '跳板机',
-        proxyCommand: '代理命令',
-        tags: '标签',
-        defaultWorkdir: '默认工作目录',
-        preferredPython: '首选 Python',
-        preferredEnvManager: '首选环境管理器',
-        preferredRuntimeNotes: '首选运行时说明',
-        sshOptionsJson: 'SSH 配置 JSON',
-        placeholders: {
-          alias: 'gpu-lab',
-          displayName: 'GPU 实验室',
-          description: '主训练环境',
-          host: 'gpu.example.com',
-          port: '22',
-          user: 'root',
-          identityFile: '/keys/gpu-lab',
-          proxyJump: 'bastion',
-          proxyCommand: 'ssh -W %h:%p bastion',
-          tags: 'gpu, research',
-          defaultWorkdir: '/workspace/project',
-          preferredPython: 'python3.13',
-          preferredEnvManager: 'uv',
-          preferredRuntimeNotes: '使用 CUDA 12 镜像',
-          sshOptionsJson: '{"StrictHostKeyChecking":"no"}',
-        },
-        authKindOptions: {
-          ssh_key: 'SSH 密钥',
-          password: '密码',
-          agent: 'SSH Agent',
-        },
-        saveError: '无法保存环境',
-        portRangeError: '端口必须介于 1 和 65535 之间',
-        noEnvironmentSelectedForEditing: '未选择要编辑的环境',
-        sshOptionsObjectError: 'SSH 配置必须是 JSON 对象',
-        sshOptionsValuesError: 'SSH 配置值必须是字符串',
-        create: '创建环境',
-        save: '保存更改',
-        saving: '保存中…',
-        reset: '重置',
-        cancelEdit: '取消编辑',
-      },
-      errorBoundary: {
-        title: '出现错误',
-        retry: '重试',
-      },
-      modal: {
-        close: '关闭',
-      },
-      skills: {
-        ungrouped: '未分组',
-        mixed: '混合',
-      },
-      token: {
-        tokens: 'Tokens',
-        total: '合计：',
-        input: '输入',
-        cache: '缓存',
-        output: '输出',
-        think: '思考',
-      },
-      resources: {
-        cpu: 'CPU 使用率',
-        cores: '核',
-        noGpu: '未检测到 GPU',
-        memory: '内存',
-      },
-      environmentDetectionModal: {
-        groups: {
-          basicInfo: '基本信息',
-          pythonToolchain: 'Python 工具链',
-          devTools: '开发工具',
-          aiMl: 'AI / ML 环境',
-          gpu: 'GPU 信息',
-          envVars: '环境变量',
-        },
-        labels: {
-          ssh: 'SSH',
-          tmux: 'Tmux',
-          hostname: '主机名',
-          os: '操作系统',
-          arch: '架构',
-          workdir: '工作目录',
-          python: 'Python',
-          conda: 'Conda',
-          uv: 'uv',
-          pixi: 'pixi',
-          codex: 'Codex',
-          claudeCli: 'Claude CLI',
-          torch: 'PyTorch',
-          cuda: 'CUDA',
-          gpuCount: 'GPU 数量',
-          gpuModels: 'GPU 型号',
-          anthropic: 'Anthropic',
-        },
-        warningsHeader: '告警',
-        status: {
-          available: '可用',
-          unavailable: '不可用',
-          present: '已配置',
-          missing: '未配置',
-          unknown: '未知',
-        },
-      },
-    },
-  },
-} as const satisfies Record<Locale, MessageTree>;
+import { messages } from './catalog';
 
+export { messages };
 export type MessageKey = Paths<(typeof messages)['en']>;
