+++ frontend/src/shared/i18n/zh/auth.ts
@@ -1,0 +1,29 @@
+// auth messages — 中文
+const auth = {
+    auth: {
+      login: '登录',
+      logout: '退出',
+      username: '用户名',
+      usernameHint: '仅限字母、数字、点、下划线、连字符',
+      password: '密码',
+      displayName: '显示名称',
+      confirmPassword: '确认密码',
+      passwordMismatch: '密码不匹配',
+      registerFailed: '注册失败',
+      registrationSubmitted: '注册已提交',
+      backToLogin: '返回登录',
+      registerLink: '没有账号？注册',
+      loginLink: '已有账号？登录',
+      register: '注册',
+      pendingApproval: '账号等待管理员审批中。',
+      changePassword: '修改密码',
+      mustChangePassword: '首次登录需要修改密码。',
+      currentPassword: '当前密码',
+      newPassword: '新密码',
+      changePasswordFailed: '修改密码失败',
+      passwordTooShort: '密码至少需要4个字符',
+      loginFailed: '登录失败',
+    },
+};
+
+export default auth;
