+++ frontend/src/shared/i18n/zh/common.ts
@@ -1,0 +1,105 @@
+// common messages — 中文
+const common = {
+    chat: {
+      thinking: '思考过程',
+      disclaimer: 'AI 可以犯错。请核实重要信息。',
+      inputPlaceholder: '给 AI 发送消息…',
+      send: '发送消息',
+      attach: '附加文件',
+      webSearch: '联网搜索',
+      reason: '深度思考',
+      copy: '复制',
+      retry: '重试',
+      collapseTools: '折叠工具',
+      scrollToBottom: '滚动到底部',
+    },
+    common: {
+      appName: 'AINRF 控制台',
+      language: '语言',
+      english: '英文',
+      chinese: '中文',
+      loading: '加载中...',
+      saving: '保存中…',
+      reset: '重置',
+      cancel: '取消',
+      save: '保存',
+      saveChanges: '保存更改',
+      createEnvironment: '创建环境',
+      use: '使用',
+      edit: '编辑',
+      detect: '探测',
+      delete: '删除',
+      default: '默认',
+      start: '启动',
+      stop: '停止',
+      active: '活动',
+      editing: '编辑中',
+      never: '从未',
+      notConfigured: '未配置',
+      notSelected: '未选择',
+      unavailable: '不可用',
+      available: '可用',
+      writable: '可写',
+      notReady: '未就绪',
+      configured: '已配置',
+      detailLabel: '详情：',
+      noTags: '无标签',
+      noDefaultWorkdir: '无默认工作目录',
+      yes: '是',
+      no: '否',
+      ok: 'OK',
+      degraded: '降级',
+      idle: '空闲',
+      starting: '启动中',
+      running: '运行中',
+      stopping: '停止中',
+      failed: '失败',
+      connecting: '连接中',
+      connected: '已连接',
+      disconnected: '已断开',
+      error: '错误',
+      builtBy: '由 Kozmosa 用 ❤️ 构建',
+      collapse: '折叠',
+      confirmLogout: '确认登出',
+      confirmLogoutMessage: '确定要登出吗？',
+      dragToReorder: '拖拽排序',
+      expand: '展开',
+      logOut: '登出',
+      taskStatusLoading: '任务 | 加载中…',
+      taskStatusSummary: '任务 | 共 {{total}} 个，运行中：{{running}}，等待中：{{pending}}，已完成：{{finished}}',
+      taskStatusUnavailable: '任务 | 状态不可用',
+      loadingEditor: '加载编辑器...',
+      noMessages: '暂无消息',
+    },
+    nav: {
+      literature: '文献追踪',
+    },
+    literature: {
+      mySubscriptions: '我的订阅',
+      newSubscription: '新建订阅',
+      label: '标签',
+      keywords: '关键词（逗号分隔）',
+      keywordsPlaceholder: '机器学习, NLP',
+      categories: '分类',
+      frequency: '频率',
+      daily: '每日',
+      weekly: '每周',
+      twice_daily: '每日两次',
+      create: '创建',
+      delete: '删除',
+      unreadOnly: '仅未读',
+      refresh: '刷新',
+      markRead: '标记已读',
+      viewarXiv: 'arXiv',
+      convertToTask: '转换为任务',
+      convertConfirm: '将此论文转换为 Agent 任务。请在下方选择工作区、运行环境和执行引擎。',
+      convertSuccess: '论文已成功转换为任务',
+      convertError: '转换论文为任务失败',
+      noPapers: '暂无论文。请先创建订阅。',
+      noSubscriptions: '暂无订阅。创建一个订阅以追踪论文。',
+      aiPracticeNote: 'AI 实践提醒',
+      aiSummary: 'AI 摘要',
+    },
+};
+
+export default common;
