+++ frontend/src/shared/i18n/zh/components.ts
@@ -1,0 +1,184 @@
+// components messages — 中文
+const components = {
+    components: {
+      healthStatusBar: {
+        loading: '正在加载系统状态...',
+        unable: '无法获取系统状态',
+        apiLabel: 'API：',
+        containerLabel: '容器：',
+        sshLabel: 'SSH：',
+        workspaceLabel: '工作区：',
+      },
+      environmentSelector: {
+        eyebrow: '环境选择器',
+        title: '共享环境选择',
+        description:
+          '只需选择一次执行环境，前端各页面都会复用它。优先级固定为设置页中的项目默认环境、本地记忆选择、localhost，最后才是列表首项。',
+        selectedNone: '未选择环境',
+        activeLabel: '当前环境',
+        noAvailable: '没有可用环境',
+        selectionStatus: '选择状态',
+        loaded: '已加载：',
+        available: '可用：',
+        loadError: '无法加载环境：',
+        displayName: '显示名称：',
+        host: '主机：',
+        defaultWorkdir: '默认工作目录：',
+        waitingValidSelection: '等待有效选择。',
+        createPrompt: '请先在“容器”中创建环境，以启用终端和工作区选择。',
+      },
+      terminalBench: {
+        eyebrow: 'PERSONAL TERMINAL SESSION',
+        title: '个人终端会话',
+        description:
+          '连接、脱离或重置当前环境的持久个人终端会话，并在浏览器中打开交互式控制台。',
+        statusPrefix: '状态：',
+        sessionSource: '会话来源',
+        loading: '加载中：',
+        websocketUrl: 'WebSocket URL:',
+        environment: '环境：',
+        sessionName: '会话名：',
+        attachment: '附着 ID：',
+        loadError: '加载错误：',
+        selectEnvironmentBeforeAttach: '附着终端会话前请先选择环境。',
+        noSessionYet: '尚未创建个人终端会话。',
+        detachedNotice: '个人会话仍在运行，但当前页面尚未附着。',
+        attach: '附着',
+        detach: '脱离',
+        resetSession: '重置会话',
+        attaching: '附着中…',
+        detaching: '脱离中…',
+        resetting: '重置中…',
+        consoleTitle: '终端控制台',
+      },
+      terminalConsole: {
+        startPrompt: '附着到个人终端会话后即可打开 xterm 控制台。',
+        writeEnabled: '可写',
+        websocketSession: 'WebSocket 会话：',
+        attachment: '附着：',
+        observeOnly: '仅观察',
+        connection: '连接：',
+        exited: '会话已退出，退出码 {{code}}',
+      },
+      environmentEditor: {
+        eyebrow: '环境编辑器',
+        createTitle: '创建环境',
+        editTitle: '编辑环境',
+        description: '定义 SSH 目标详情、运行时偏好，以及支撑环境控制平面的最小元数据。',
+        activeLabel: '当前活动环境：',
+        noActive: '尚未标记任何活动环境。',
+        alias: '别名',
+        displayName: '显示名称',
+        descriptionField: '描述',
+        host: '主机',
+        port: '端口',
+        user: '用户',
+        authKindLabel: '认证方式',
+        identityFile: '身份文件',
+        proxyJump: '跳板机',
+        proxyCommand: '代理命令',
+        tags: '标签',
+        defaultWorkdir: '默认工作目录',
+        preferredPython: '首选 Python',
+        preferredEnvManager: '首选环境管理器',
+        preferredRuntimeNotes: '首选运行时说明',
+        sshOptionsJson: 'SSH 配置 JSON',
+        placeholders: {
+          alias: 'gpu-lab',
+          displayName: 'GPU 实验室',
+          description: '主训练环境',
+          host: 'gpu.example.com',
+          port: '22',
+          user: 'root',
+          identityFile: '/keys/gpu-lab',
+          proxyJump: 'bastion',
+          proxyCommand: 'ssh -W %h:%p bastion',
+          tags: 'gpu, research',
+          defaultWorkdir: '/workspace/project',
+          preferredPython: 'python3.13',
+          preferredEnvManager: 'uv',
+          preferredRuntimeNotes: '使用 CUDA 12 镜像',
+          sshOptionsJson: '{"StrictHostKeyChecking":"no"}',
+        },
+        authKindOptions: {
+          ssh_key: 'SSH 密钥',
+          password: '密码',
+          agent: 'SSH Agent',
+        },
+        saveError: '无法保存环境',
+        portRangeError: '端口必须介于 1 和 65535 之间',
+        noEnvironmentSelectedForEditing: '未选择要编辑的环境',
+        sshOptionsObjectError: 'SSH 配置必须是 JSON 对象',
+        sshOptionsValuesError: 'SSH 配置值必须是字符串',
+        create: '创建环境',
+        save: '保存更改',
+        saving: '保存中…',
+        reset: '重置',
+        cancelEdit: '取消编辑',
+      },
+      errorBoundary: {
+        title: '出现错误',
+        retry: '重试',
+      },
+      modal: {
+        close: '关闭',
+      },
+      skills: {
+        ungrouped: '未分组',
+        mixed: '混合',
+      },
+      token: {
+        tokens: 'Tokens',
+        total: '合计：',
+        input: '输入',
+        cache: '缓存',
+        output: '输出',
+        think: '思考',
+      },
+      resources: {
+        cpu: 'CPU 使用率',
+        cores: '核',
+        noGpu: '未检测到 GPU',
+        memory: '内存',
+      },
+      environmentDetectionModal: {
+        groups: {
+          basicInfo: '基本信息',
+          pythonToolchain: 'Python 工具链',
+          devTools: '开发工具',
+          aiMl: 'AI / ML 环境',
+          gpu: 'GPU 信息',
+          envVars: '环境变量',
+        },
+        labels: {
+          ssh: 'SSH',
+          tmux: 'Tmux',
+          hostname: '主机名',
+          os: '操作系统',
+          arch: '架构',
+          workdir: '工作目录',
+          python: 'Python',
+          conda: 'Conda',
+          uv: 'uv',
+          pixi: 'pixi',
+          codex: 'Codex',
+          claudeCli: 'Claude CLI',
+          torch: 'PyTorch',
+          cuda: 'CUDA',
+          gpuCount: 'GPU 数量',
+          gpuModels: 'GPU 型号',
+          anthropic: 'Anthropic',
+        },
+        warningsHeader: '告警',
+        status: {
+          available: '可用',
+          unavailable: '不可用',
+          present: '已配置',
+          missing: '未配置',
+          unknown: '未知',
+        },
+      },
+    },
+};
+
+export default components;
