+++ frontend/src/shared/i18n/zh/navigation.ts
@@ -1,0 +1,56 @@
+// navigation messages — 中文
+const navigation = {
+    navigation: {
+      projects: {
+        label: '项目',
+        description: '项目画布与任务图可视化',
+      },
+      terminal: {
+        label: '终端',
+        description: '健康检查和终端工作台入口',
+      },
+      tasks: {
+        label: '任务',
+        description: 'Agent 窗口与只读观察终端',
+      },
+      workspaces: {
+        label: '工作区',
+        description: '项目工作区和任务上下文',
+      },
+      workspaceBrowser: {
+        label: '工作区浏览',
+        description: '浏览工作区文件',
+      },
+      environments: {
+        label: '容器',
+        description: '运行环境与容器状态',
+      },
+      resources: {
+        label: '资源监控',
+        description: '系统资源占用与使用情况',
+      },
+      sessions: {
+        label: '任务运行',
+        description: '来自真实 Agent 任务的运行历史',
+      },
+      timeline: {
+        label: '任务时间线',
+        description: '来自真实 Agent 任务的运行时间线',
+      },
+      settings: {
+        label: '设置',
+        description: '运行时和 WebUI 偏好设置',
+      },
+    },
+    layout: {
+      brandLine: '与后端健康检查、终端工作台、环境控制平面和托管工作区浏览器对齐的轻量前端壳。',
+      headerEyebrow: 'WebUI 控制面',
+      headerDescription: '后端重整期间的健康、终端工作台、环境控制平面和工作区浏览器外壳。',
+      footerDescription: '后端重整期间的健康、终端工作台、环境控制平面和工作区浏览器外壳。',
+      collapseSidebar: '收起侧边栏',
+      expandSidebar: '展开侧边栏',
+      resizeSidebar: '调整侧边栏宽度',
+    },
+};
+
+export default navigation;
