+++ frontend/src/shared/i18n/zh/pages.ts
@@ -1,0 +1,709 @@
+// pages messages — 中文
+const pages = {
+    pages: {
+      projects: {
+        sidebarEyebrow: '项目',
+        sidebarTitle: '项目画布',
+        sidebarCount: '{{count}} 个项目',
+        newProject: '新建项目',
+        searchPlaceholder: '搜索项目',
+        newTask: '新建任务',
+        resetLayout: '重置布局',
+        emptyCanvas: "点击'新建任务'开始",
+        noProjects: '暂无项目。创建一个项目开始。',
+        moveToProject: '移动到项目',
+        noOtherProjects: '没有其他可移动到的项目。',
+        createTitle: '创建项目',
+        createNameLabel: '项目名称',
+        createDescriptionLabel: '描述（可选）',
+        createSubmit: '创建项目',
+      },
+      dashboard: {
+        eyebrow: '终端',
+        title: 'Scholar Agent 运行时外壳',
+        description:
+          '前端现在暴露了整理后的运行时外壳：健康检查、单会话终端工作台控制面、环境控制平面和托管工作区浏览面板。此页用于确认 API、SSH、工作区就绪状态、终端会话状态和环境就绪状态。',
+        surfaceTitle: '当前后端面',
+        endpointsLabel: '可用端点：',
+        stateRootLabel: '状态根：',
+        detailLabel: '详情：',
+      },
+      terminal: {
+        eyebrow: 'TERMINAL',
+        title: '终端',
+        description: '',
+        backToTasks: '返回任务页',
+        task: {
+          observeOnly: '仅观察',
+          takenOver: '已接管',
+          archived: '已归档',
+          pending: '等待附着',
+          windowLabel: '窗口：',
+          ownerLabel: '所有者：',
+          environmentLabel: '环境：',
+          noOwner: '当前无人占用',
+          lastOutputLabel: '最近输出：',
+          otherOwner: '该任务当前由其他浏览器用户接管，此处无法释放。',
+          placeholder: '任务终端附着尚未激活。',
+        },
+      },
+      tasks: {
+        sidebarEyebrow: '任务运行',
+        sidebarTitle: 'Agent 任务',
+        sidebarCount: '共 {{count}} 个 · 实时输出回放',
+        newTask: '新建任务',
+        resizeTaskList: '调整任务列表宽度',
+        searchLabel: '搜索任务',
+        searchPlaceholder: '搜索任务',
+        empty: '还没有任务。新建一个运行后即可开始。',
+        noSearchResults: '没有匹配“{{query}}”的任务。',
+        updatedAt: '更新于 {{time}}',
+        createTitle: '创建任务',
+        createDescription: '在工作区与环境绑定上启动一个 Claude Code 任务。',
+        closeCreate: '关闭创建任务',
+        workspaceLabel: '工作区',
+        environmentLabel: '环境',
+        projectLabel: '项目',
+        profileLabel: '任务配置',
+        researchAgentLabel: '研究代理',
+        taskConfigurationLabel: '任务配置模式',
+        skillsLabel: '技能',
+        skillDependenciesHint: '有依赖的技能将在运行时自动包含所需技能。',
+        titleLabel: '标题',
+        optionalPlaceholder: '可选',
+        taskInputLabel: '任务输入',
+        taskInputPlaceholder: '描述要交给 Agent 的任务。',
+        structuredResearchGoal: '研究目标',
+        structuredResearchContext: '上下文',
+        structuredResearchConstraints: '约束',
+        structuredResearchDeliverables: '预期交付物',
+        structuredResearchValidationPlan: '验证计划',
+        generatedPromptPreview: '生成提示词预览',
+        paperPathLabel: '论文路径',
+        paperPathPlaceholder: 'papers/target.pdf',
+        scopeLabel: '范围',
+        targetTableLabel: '目标表格（可选）',
+        targetTablePlaceholder: 'Table3',
+        budgetHoursLabel: '预算（小时）',
+        researchTopicLabel: '研究主题',
+        researchTopicPlaceholder: '用于药物发现的图神经网络',
+        seedPaperPathLabel: '种子论文路径（可选）',
+        seedPaperPathPlaceholder: 'papers/seed.pdf',
+        depthLabel: '深度',
+        ideaSourceLabel: '想法来源',
+        ideaSourcePlaceholder: '工作区中的想法文档路径，或在此描述你的想法',
+        validationScopeLabel: '验证范围',
+        scopeCoreOnly: '核心方法',
+        scopeFullSuite: '完整套件',
+        validationScopeQuick: '快速',
+        validationScopeFull: '完整',
+        defaultWorkdir: '默认工作目录：',
+        resetDraft: '重置草稿',
+        createAction: '创建任务',
+        creatingAction: '创建中…',
+        create: {
+          taskPreset: '任务预设',
+          presets: {
+            rawPrompt: '原始提示词',
+            structuredResearch: '结构化研究',
+            reproduceBaseline: '复现实验基线',
+            overview: '概要总结',
+          },
+          researcherType: '研究员类型',
+          researcherVanilla: 'Vanilla',
+          researcherAris: 'ARIS Researcher',
+          executionEngine: '执行引擎',
+          engineClaudeCode: 'Claude Code',
+          engineAgentSdk: 'Agent SDK',
+          engineCodexAppServer: 'Codex App Server',
+          skillsPlaceholder: 'analysis, code-review',
+          titlePlaceholder: '可选任务标题',
+          promptLabel: 'Prompt',
+          promptPlaceholder: '输入研究任务提示词…',
+          missingBinding: '请先选择工作区和环境再创建任务。',
+          noSkillsAvailable: '当前没有可用于此任务的技能。',
+          skillGroupSelected: '已选择 {{selected}}/{{total}}',
+          showSkillGroup: '显示 {{group}} 中的技能',
+          hideSkillGroup: '隐藏 {{group}} 中的技能',
+          selectSkill: '选择 {{skill}}',
+          deselectSkill: '取消选择 {{skill}}',
+          seedFiles: {
+            dropzone: '拖拽种子论文到此处，或点击选择文件（支持 PDF、Markdown）',
+            uploading: '正在上传 {{name}}…',
+            uploadError: '上传 {{name}} 失败：{{error}}',
+            removeFile: '移除 {{name}}',
+            unsupportedType: '仅支持 PDF 和 Markdown 文件。',
+            maxFilesExceeded: '最多上传 {{max}} 篇种子论文。',
+            fileCount: '已上传 {{count}} 篇种子论文',
+            label: '种子论文',
+          },
+        },
+        noTaskSelected: '未选择任务',
+        noTaskSelectedDescription: '从侧边栏选择一个任务，或新建一次运行来查看输出。',
+        workspaceEyebrow: '任务工作区',
+        outputTimeline: '输出时间线',
+        latestSeq: '最新序号 {{seq}}',
+        noOutput: '尚未记录任何输出。',
+        promptLayers: 'Prompt 分层',
+        prompt: 'Prompt',
+        chars: '字符',
+        promptUnavailable: 'Prompt 清单尚不可用。',
+        summary: '摘要',
+        taskId: '任务 ID',
+        created: '创建时间',
+        updated: '更新时间',
+        started: '开始时间',
+        completed: '完成时间',
+        binding: '绑定',
+        profile: '配置',
+        workdir: '工作目录',
+        snapshot: '快照',
+        bindingUnavailable: '绑定快照尚不可用。',
+        runtime: '运行时',
+        runner: 'Runner',
+        directory: '目录',
+        command: '命令',
+        runtimeUnavailable: '运行时载荷尚不可用。',
+        result: '结果',
+        exitCode: '退出码',
+        failure: '失败',
+        unavailable: '不可用',
+        pendingValue: '等待中',
+        status: {
+          queued: '排队中',
+          starting: '启动中',
+          running: '运行中',
+          succeeded: '已成功',
+          failed: '已失败',
+          cancelled: '已取消',
+          paused: '已暂停',
+        },
+        expandOutputTimeline: '展开输出时间线',
+        collapseOutputTimeline: '收起输出时间线',
+        expandSummary: '展开摘要面板',
+        collapseSummary: '收起摘要面板',
+        thinking: '思考中…',
+        toolCall: '工具：{{name}}',
+        toolResult: '结果',
+        inputPlaceholder: '输入消息…',
+        metadata: {
+          workdir: '工作目录',
+          command: '命令',
+          taskInput: '任务输入',
+        },
+        layout: {
+          showOnlyDetails: '仅显示详情',
+          showOnlyConversation: '仅显示对话',
+          collapseSidebar: '收起详情',
+          expandSidebar: '展开详情',
+          resizeSidebar: '调整详情面板宽度',
+          errorIndicator: '任务有错误 — 点击查看',
+        },
+        output: {
+          replayFailed: '无法回放任务输出',
+          parseFailed: '无法解析任务输出',
+          loadFailed: '无法加载任务输出',
+        },
+        messageGroup: {
+          summary: '{{count}} 条{{type}}消息',
+        },
+        messageType: {
+          systemEvent: '系统',
+          user: '用户',
+          assistant: '助手',
+          thinking: '思考',
+          toolCall: '工具调用',
+          toolResult: '工具结果',
+          stdout: '标准输出',
+          stderr: '标准错误',
+          lifecycle: '生命周期',
+        },
+        actions: {
+          cancel: '取消',
+          archive: '归档',
+          retry: '重试',
+          showArchived: '显示已归档',
+          pause: '暂停',
+          resume: '继续',
+          pauseFailed: '暂停失败：{{error}}',
+          resumeFailed: '继续失败：{{error}}',
+          sendPromptFailed: '发送提示词失败：{{error}}',
+          noTaskSelected: '未选择任务',
+          unexpectedError: '发生意外错误',
+        },
+        retrySuccess: '任务已重试',
+        retryFailed: '重试失败',
+        sort: {
+          updated: '更新时间（从新到旧）',
+          created: '创建时间（从新到旧）',
+          name: '名称（A-Z）',
+        },
+      },
+      workspaces: {
+        eyebrow: 'WORKSPACES',
+        title: '工作区',
+        description:
+          '管理可复用的工作区上下文，用于任务提示词、默认工作目录和研究运行边界。',
+        managerTitle: '工作区管理器',
+        managerDescription: '创建和维护任务启动前可绑定的工作区记录。',
+        newWorkspace: '新建工作区',
+        createWorkspace: '创建工作区',
+        saveWorkspace: '保存工作区',
+        deleteWorkspace: '删除工作区',
+        confirmDelete: '确认删除',
+        labelField: '工作区名称',
+        descriptionField: '说明',
+        defaultWorkdirField: '默认工作目录',
+        promptField: '工作区提示词',
+        noDefaultWorkdir: '未配置默认工作目录',
+      },
+      workspaceBrowser: {
+        eyebrow: 'WORKSPACE BROWSER',
+        title: '工作区浏览器',
+        description: '打开设置页所选环境的托管 code-server 工作区浏览器。',
+        readinessTitle: '需要完成运行时设置',
+        readinessMissingDependency: '缺少运行时依赖。',
+      },
+      settings: {
+        eyebrow: 'SETTINGS',
+        title: '设置',
+        description:
+          '自定义 Scholar Agent 的使用体验：默认首页、终端字号、项目默认环境以及任务草稿模板。',
+        recoveryNotice:
+          '本地设置文档缺字段、已损坏或版本不兼容，应用已回退到默认配置以继续运行。',
+        routes: {
+          terminal: '终端',
+          tasks: '任务',
+          workspaces: '工作区',
+          environments: '环境',
+        },
+        general: {
+          title: '通用偏好',
+          description: '作用于整个界面的偏好设置。',
+          defaultRouteLabel: '默认首页',
+          terminalFontSizeLabel: '终端字号',
+          terminalFontSizeHelp:
+            '字号会被限制在稳定区间 {{min}}-{{max}} px，当前保存值为 {{current}} px。',
+          editorFontSizeLabel: '编辑器字号',
+          editorFontFamilyLabel: '编辑器字体',
+          editorFontSizeHelp:
+            '编辑器字号会被限制在 {{min}}-{{max}} px，当前保存值为 {{current}} px。',
+        },
+        project: {
+          title: '项目默认值',
+          description: '当前项目的默认配置，仅保存在当前浏览器中。',
+          defaultEnvironmentLabel: '默认环境',
+          defaultEnvironmentEmpty: '不设置项目默认环境',
+          defaultEnvironmentHelp:
+            '这个环境优先级高于本地记忆的最近一次显式选择。',
+          noEnvironments:
+            '请先在“容器”页面创建至少一个环境，再配置项目默认值。',
+          environmentCardDescription: '这些字段只用于预填当前环境下的新建任务草稿。',
+          titleTemplateLabel: '标题模板',
+          titleTemplatePlaceholder: '可选的任务标题预设',
+          taskInputTemplateLabel: '任务输入模板',
+          taskInputTemplatePlaceholder: '填写该环境下默认预载入的任务输入。',
+          researchAgentDefaultLabel: '默认研究代理',
+          taskConfigurationDefaultLabel: '默认任务配置',
+        },
+        taskConfiguration: {
+          title: '任务配置',
+          description:
+            '配置创建任务时的三层结构：Claude Code 执行引擎、Research Agent 配置和任务输入模式。',
+          executionEngineLabel: '执行引擎',
+          defaultResearchAgentLabel: '默认研究代理',
+          defaultTaskConfigurationLabel: '默认任务配置',
+          profileLabel: 'Research Agent 标签',
+          systemPromptLabel: '系统提示词',
+          skillsLabel: '启用的技能',
+          skillsDescription: '选择该研究代理应使用的技能。',
+          skillsPromptLabel: 'Skills / 配置说明',
+          settingsJsonLabel: 'Claude settings JSON',
+          apiBaseUrlLabel: 'API Base URL',
+          apiKeyLabel: 'API Key',
+          defaultOpusModelLabel: 'Opus 模型',
+          defaultSonnetModelLabel: 'Sonnet 模型',
+          defaultHaikuModelLabel: 'Haiku 模型',
+          envOverridesLabel: '环境变量覆盖 (JSON)',
+          placeholders: {
+            apiBaseUrl: 'https://api.anthropic.com/',
+            apiKey: 'sk-...',
+            opusModel: 'claude-opus-4-7',
+            sonnetModel: 'claude-sonnet-4-6',
+            haikuModel: 'claude-haiku-4-5',
+            envOverrides: '{\n  "CUSTOM_VAR": "value"\n}',
+          },
+        },
+        skillRepository: {
+          title: 'Skill 仓库',
+          description: '管理 AINRF 技能：查看详情、预览设置并导入新技能。',
+          selectSkill: '选择一个技能以查看详情',
+          version: '版本',
+          author: '作者',
+          dependencies: '依赖',
+          mcpServers: 'MCP 服务器',
+          hooks: '钩子',
+          allowedAgents: '允许的 Agent',
+          skillMdTitle: 'SKILL.md',
+          previewSettings: '预览设置',
+          settingsPreviewTitle: '设置预览',
+          importSkill: '导入技能',
+          importTitle: '导入技能',
+          sourceLabel: '来源',
+          gitSource: 'Git URL',
+          localSource: '本地路径',
+          urlLabel: 'Git URL',
+          pathLabel: '本地路径',
+          skillIdOverrideLabel: 'Skill ID 覆盖（可选）',
+          importAction: '导入',
+          importing: '导入中…',
+          cancel: '取消',
+          noSkills: '仓库中未找到任何技能。',
+          placeholders: {
+            url: 'https://github.com/example/skill-repo.git',
+            path: '/path/to/skill',
+          },
+        },
+        defaultWorkspace: {
+          title: '默认工作区',
+          description: '选择默认工作区用于任务创建和文件浏览。',
+          label: '默认工作区',
+          noDefault: '无默认工作区',
+        },
+        searchBackend: {
+          title: '网络搜索后端',
+          description: '选择 Agent 会话使用的网页搜索和内容抓取工具。',
+          activeBackend: '活跃后端',
+          autoStart: '自动启动 MCP 服务器',
+          autoStartHint: '选中的 MCP 服务器会在每个任务会话中自动启动。',
+          native: 'Claude 原生',
+          kindlyWebSearch: 'Kindly Web Search',
+          ccWebMcp: 'CC-Web-MCP',
+          saving: '保存中…',
+          saved: '已保存',
+        },
+        version: {
+          title: '部署版本',
+          description: '当前运行的后端与前端构建各自内置的 Git 提交。',
+          backendLabel: '后端',
+          frontendLabel: '前端',
+          commitLabel: '提交',
+          committedAtLabel: '提交时间',
+          unavailable: '不可用',
+          mismatchWarning: '后端与前端基于不同的提交构建，版本不一致。',
+        },
+        account: {
+          title: '账号与安全',
+          description: '管理你的账号凭证。',
+        },
+        tabs: {
+          general: '通用',
+          llmProviders: 'LLM 提供商',
+          monitoring: '监控审计',
+          users: '用户管理',
+          envAccess: '环境访问',
+          collaborators: '协作者',
+        },
+        monitoring: {
+          title: '可观测性平台',
+          description: '快速访问监控、指标和 LLM 可观测性仪表盘。',
+          noServices: '未配置任何监控服务。请在后端设置相应的环境变量以启用链接。',
+          notConfigured: '未配置',
+          openService: '打开 {{name}}',
+          grafanaDesc: '指标仪表盘、告警与数据可视化',
+          prometheusDesc: '时序指标采集与查询',
+          litefuseDesc: 'LLM 可观测性：链路追踪、生成记录与 Token 分析',
+        },
+        envAccess: {
+          selectEnv: '选择环境...',
+          maxTasks: '最大任务数',
+          unlimited: '无限制',
+          grant: '授权',
+          remove: '移除',
+          grantTo: '授权给...',
+          noEnvSelected: '选择一个环境以管理访问权限。',
+          noAccess: '暂无授权记录。',
+        },
+        collaborators: {
+          selectProject: '选择项目...',
+          add: '添加',
+          addUser: '添加用户...',
+          remove: '移除',
+          role: {
+            member: '成员',
+            viewer: '观察者',
+          },
+          noProjectSelected: '选择一个项目以管理协作者。',
+          noCollaborators: '暂无协作者。',
+          loadError: '加载协作者失败',
+        },
+        engine: {
+          claudeCode: 'Claude Code',
+          claudeAgent: 'Agent SDK',
+          codexAppServer: 'Codex App Server',
+        },
+        codex: {
+          model: '模型',
+          command: '命令',
+          approval: '审批',
+          baseUrl: 'Base URL',
+          apiKey: 'API Key',
+          config: 'Codex 配置 (TOML)',
+          auth: 'Codex 认证 (JSON)',
+          install: '安装 {{name}}',
+          installing: '正在安装 {{name}}…',
+          update: '更新 {{name}}',
+          updating: '更新中…',
+          installed: '{{name}} 已安装',
+          updateTitle: '更新 {{name}}',
+          updateWarning: '此操作将覆盖现有文件，是否继续？',
+          forceUpdate: '强制更新',
+          placeholders: {
+            baseUrl: 'https://api.openai.com/',
+            apiKey: 'sk-...',
+            model: 'gpt-5-codex',
+            command: 'codex app-server --listen stdio://',
+            approval: 'never',
+            config: 'model = "gpt-5-codex"',
+            auth: '{\n  "OPENAI_API_KEY": "sk-..."\n}',
+          },
+        },
+        appearance: {
+          title: '外观',
+          description: '全局字体偏好设置',
+          fontFamilyLabel: '字体',
+          sansSerif: '非衬线体',
+          serif: '衬线体',
+          previewHint: '更改将立即应用到整个界面。',
+        },
+        users: {
+          title: '用户管理',
+          count: '{{count}} 个用户',
+          approve: '审批通过',
+          disable: '禁用',
+          reEnable: '重新启用',
+          resetPassword: '重置密码',
+          newPassword: '新密码：',
+          enterNewPassword: '输入新密码',
+          set: '设置',
+        },
+        llmProviders: {
+          title: 'LLM 提供商',
+          description: '管理全局 LLM API 配置，以便在 Agent 配置中快速填充。',
+          addProvider: '添加提供商',
+          editProvider: '编辑提供商',
+          deleteProvider: '删除提供商',
+          noProviders: '尚未配置任何提供商。添加一个以在 Agent 配置中启用快速填充。',
+          nameLabel: '名称',
+          formatLabel: '格式',
+          baseUrlLabel: 'Base URL',
+          apiKeyLabel: 'API Key',
+          opusModelLabel: 'Opus 模型',
+          sonnetModelLabel: 'Sonnet 模型',
+          haikuModelLabel: 'Haiku 模型',
+          defaultModelLabel: '默认模型',
+          formatAnthropic: 'Anthropic',
+          formatOpenAIChat: 'OpenAI Chat',
+          formatOpenAIResponses: 'OpenAI Responses',
+          confirmDelete: '删除提供商「{{name}}」？',
+          fillFromProvider: '从提供商填充',
+          customOption: '自定义',
+          placeholders: {
+            name: '例如 Kimi Coding',
+            anthropicBaseUrl: 'https://api.anthropic.com/',
+            openAiBaseUrl: 'https://api.openai.com/',
+            apiKey: 'sk-...',
+            opusModel: 'claude-opus-4-7',
+            sonnetModel: 'claude-sonnet-4-6',
+            haikuModel: 'claude-haiku-4-5',
+            defaultModel: 'gpt-4o',
+          },
+          updateFailed: '更新失败',
+        },
+      },
+      placeholder: {
+        eyebrow: '设置',
+        title: '设置即将推出',
+        description: '此区域将在后续更新中支持环境偏好、运行时配置和行为控制。',
+        badge: '占位区块',
+      },
+      environments: {
+        eyebrow: 'CONTAINERS',
+        title: '容器',
+        description:
+          '添加并管理基于 SSH 的计算环境。运行连通性探测，然后将其中一个标记为活动环境，供终端和任务使用。',
+        currentSelection: '当前选择',
+        activeSelectionFallback: '尚未选择活动环境',
+        addEnvironment: '添加环境',
+        listTitle: '环境列表',
+        listDescription: '选择、编辑、探测或删除环境。活动环境会被高亮显示。',
+        loading: '正在加载环境...',
+        empty: '尚未创建任何环境。',
+        alias: '别名',
+        host: '主机',
+        auth: '认证',
+        detection: '探测',
+        actions: '操作',
+        defaultWorkdir: '未配置默认工作目录',
+        activeBadge: '活动',
+        editingBadge: '编辑中',
+        detectedAt: '探测于',
+        notDetected: '尚未探测',
+        confirmDelete: '删除环境「{{alias}}」？',
+        defaultEnvironmentLocked: '默认的 localhost 环境不能删除。',
+        detectFallbackToast: '{{alias}} 探测在 SSH 不可用后已回退到个人 tmux 会话。',
+        detectFailedToast: '环境探测失败。',
+        detectionStatus: {
+          success: '成功',
+          partial: '部分成功',
+          failed: '失败',
+        },
+        authKind: {
+          ssh_key: 'SSH 密钥',
+          password: '密码',
+          agent: 'SSH Agent',
+        },
+        projectReferenceTitle: '项目环境绑定',
+        projectReferenceDescription:
+          '把当前所选环境绑定到当前项目，设置项目默认环境，并覆盖仅与运行时相关的字段，例如工作目录或环境管理器。',
+        projectReferenceNoSelection: '请先从列表中选择一个环境，再编辑项目绑定。',
+        projectReferencedBadge: '项目引用',
+        projectUnreferencedBadge: '未引用',
+        projectDefaultBadge: '项目默认',
+        setProjectDefault: '默认',
+        setProjectDefaultDisabled: '已是默认',
+        useTooltip: '用作会话默认',
+        defaultTooltip: '设为新项目默认',
+        clearProjectDefault: '取消项目默认',
+        removeProjectReference: '移除项目引用',
+        projectReferenceSaveError: '无法保存项目环境绑定',
+        projectOverrideWorkdir: '覆盖工作目录',
+        projectOverrideWorkdirPlaceholder: '/workspace/project-override',
+        projectOverrideEnvName: '覆盖环境名',
+        projectOverrideEnvNamePlaceholder: 'research-env',
+        projectOverrideEnvManager: '覆盖环境管理器',
+        projectOverrideEnvManagerPlaceholder: 'conda',
+        projectOverrideRuntimeNotes: '覆盖运行时说明',
+        projectOverrideRuntimeNotesPlaceholder: '优先使用项目专用运行时镜像。',
+        projectReferenceSaving: '正在保存绑定…',
+        updateProjectReference: '更新项目绑定',
+        attachProjectReference: '绑定到项目',
+      },
+      resources: {
+        eyebrow: '资源监控',
+        title: '资源监控',
+        description: '所有环境的实时系统资源占用情况。',
+        loading: '正在加载资源数据...',
+        noData: '暂无资源数据。',
+        systemCard: {
+          title: '系统资源',
+          gpuTitle: 'GPU',
+          cpuTitle: 'CPU',
+          memoryTitle: '内存',
+          noGpu: '未检测到 GPU',
+        },
+        processCard: {
+          title: 'AINRF 进程',
+          empty: '未检测到 AINRF 进程',
+          columns: {
+            pid: 'PID',
+            name: '进程名',
+            cpu: 'CPU%',
+            memory: '内存',
+            runtime: '运行时间',
+          },
+        },
+        taskUsage: {
+          title: '任务用量',
+          description: '所有任务的 token 与耗时统计。',
+          taskCount: '{{count}} 个任务',
+          totalTokens: '总 Token',
+          totalDuration: '任务总耗时',
+          medianDuration: '任务耗时中位数',
+          topTasks: 'Token 消耗最高的任务',
+          tokens: '{{count}} 个 token',
+          noTopTasks: '暂无任务 token 用量记录。',
+        },
+        status: {
+          ok: '正常',
+          degraded: '降级',
+          unavailable: '不可用',
+        },
+      },
+      sessions: {
+        sidebarTitle: '任务运行',
+        sidebarCount: '{{count}} 个运行',
+        searchPlaceholder: '搜索任务运行...',
+        empty: '暂无任务运行',
+        selectPrompt: '选择一个任务运行查看详情',
+        notFound: '未找到任务运行',
+        outputCount: '{{count}} 条输出事件',
+        taskCount: '{{count}} 个任务',
+        attemptsTitle: '尝试记录',
+        noAttempts: '暂无尝试记录',
+        attemptLabel: '尝试 #{{seq}}',
+        viewTask: '查看任务',
+        hasTokens: '有 Token 数据',
+        status: {
+          active: '活跃',
+          completed: '已完成',
+          archived: '已归档',
+        },
+        attemptStatus: {
+          running: '运行中',
+          completed: '已完成',
+          failed: '失败',
+          interrupted: '已中断',
+        },
+        perModelBreakdown: '模型明细',
+        detail: {
+          prompt: '提示词',
+          engine: '引擎',
+          project: '项目',
+          workspace: '工作区',
+          environment: '环境',
+          workdir: '工作目录',
+          command: '命令',
+          started: '开始',
+          completed: '完成',
+          outputSeq: '输出事件',
+          exitCode: '退出码',
+        },
+        totalTokens: '{{count}} 个 token',
+        fileBrowser: {
+          loading: '加载文件...',
+          files: '文件',
+          noFileSelected: '未选择文件',
+          noFiles: '无文件',
+          loadingFile: '加载文件...',
+          selectEnv: '选择一个环境以浏览文件',
+          refresh: '刷新',
+          selectFile: '选择一个文件以查看内容',
+          binaryFile: '二进制文件',
+          emptyDirectory: '空目录',
+        },
+        timeline: {
+          title: '任务运行时间线',
+          empty: '此时间段内无任务运行',
+          loading: '正在加载任务时间线...',
+          taskCount: '{{count}} 个任务运行',
+          projectFilter: '按项目筛选任务时间线',
+          allProjects: '全部项目',
+          today: '今天',
+          past7Days: '过去 7 天',
+          past30Days: '过去 30 天',
+          from: '从',
+          to: '至',
+          attemptSeq: '第{seq}次',
+          status: '状态：',
+          duration: '耗时：',
+          reason: '原因：',
+          attemptLabel: '尝试 #{{seq}}',
+          viewTask: '查看任务',
+          viewSession: '查看会话',
+        },
+      },
+    },
+};
+
+export default pages;
